Resample Callback

Specialized mlr3misc::Callback to customize the behavior of resample() and benchmark() in mlr3. For example, callbacks can be used to extract information from models on the worker or to store intermediate results to disk. The callback_resample() function is used to create instances of this class. Predefined callbacks are stored in the dictionary mlr_callbacks and can be retrieved with clbk(). For more information on callbacks, see the callback_resample() documentation.

Super class

mlr3misc::Callback -> CallbackResample

Public fields

on_resample_begin

(function())
Stage called at the beginning of the resampling iteration. Called in workhorse() (internal).

on_resample_before_train

(function())
Stage called before training the learner. Called in workhorse() (internal).

on_resample_before_predict

(function())
Stage called before predicting. Called in workhorse() (internal).

on_resample_end

(function())
Stage called at the end of the resample iteration. Called in workhorse() (internal).

Methods

Public methods

CallbackResample$clone()

Inherited methods

mlr3misc::Callback$call()

mlr3misc::Callback$format()

mlr3misc::Callback$help()

mlr3misc::Callback$initialize()

mlr3misc::Callback$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

CallbackResample$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.