DataBackend for Matrix
Source: R/DataBackendMatrix.R

DataBackend for Matrix. Data is stored as (sparse) matrix.

Format

R6::R6Class object inheriting from DataBackend.

Construction

DataBackendMatrix$new(data, primary_key = NULL)

data :: Matrix::Matrix().

primary_key :: character(1)
Not supported by this backend. Rows are addresses by their rownames(). If the matrix does not have row names, integer row indices are used.

Alternatively, use as_data_backend on a Matrix::Matrix().

Fields

nrow :: integer(1)
Number of rows (observations).

ncol :: integer(1)
Number of columns (variables), including the primary key column.

colnames :: character()
Returns vector of all column names, including the primary key column.

rownames :: integer() | character()
Returns vector of all distinct row identifiers, i.e. the primary key column.

Methods

data(rows = NULL, cols = NULL, format = "data.table")
(integer() | character(), character()) -> any
Returns a slice of the data in a specific format. Currently, the only supported format is "data.table". The rows must be addressed as vector of primary key values, columns must be referred to via column names. Non-existing rows and columns are silently ignored.

distinct(cols)
character() -> named list()
Returns a named list of vectors of distinct values for each column specified. Non-existing columns are silently ignored.

head(n = 6)
integer(1) -> data.table::data.table()
Returns the first up-to n rows of the data as data.table::data.table().

missing(rows, cols)
(integer() | character(), character()) -> named integer()
Returns the number of missing values per column in the specified slice of data. Non-existing rows and columns are silently ignored.

See also

Examples

requireNamespace("Matrix")

#> Loading required namespace: Matrix

data = Matrix::Matrix(sample(0:1, 20, replace = TRUE), ncol = 2)
colnames(data) = c("x1", "x2")
rownames(data) = paste0("row_", 1:10)

b = as_data_backend(data)
print(b$head(n = 3, format = "data.table"))

#>    x1 x2 ..row_id
#> 1:  0  1    row_1
#> 2:  1  1    row_2
#> 3:  0  0    row_3

print(b$head(n = 3, format = "sparse"))

#> 3 x 2 sparse Matrix of class "dgCMatrix"
#>       x1 x2
#> row_1  .  1
#> row_2  1  1
#> row_3  .  .