Abstract Prediction Object
Source: R/Prediction.R

This is the abstract base class for task objects like PredictionClassif or PredictionRegr.

Prediction objects store the following information:

The row ids of the test set

The corresponding true (observed) response.

The corresponding predicted response.

Additional predictions based on the class and predict_type. E.g., the class probabilities for classification or the estimated standard error for regression.

Format

R6::R6Class object.

Construction

This object is constructed via a derived classes, e.g. PredictionClassif or PredictionRegr.

Fields

row_ids :: (integer() | character())
Vector of row ids for which predictions are stored.

truth :: any
True (observed) outcome.

task_type :: character(1)
Stores the type of the Task.

predict_types :: character()
Vector of predict types this object stores.

missing :: logical()
Returns row_ids for which the predictions are missing or incomplete.

Methods

score(measures = NULL, task = NULL, learner = NULL)
(list of Measure, Task, Learner) -> Prediction
Calculates the performance for all provided measures Task and Learner may be NULL for most measures, but some measures need to extract information from these objects.

S3 Methods

as.data.table(rr)
Prediction -> data.table::data.table()
Converts the data to a data.table::data.table().

c(..., keep_duplicates = TRUE)
(Prediction, Prediction, ...) -> Prediction
Combines multiple Predictions to a single Prediction. If keep_duplicates is FALSE and there are duplicated row ids, the data of the former passed objects get overwritten by the data of the later passed objects.

See also