Resampling Class
Source: R/Resampling.R

This is the abstract base class for resampling objects like ResamplingCV and ResamplingBootstrap.

The objects of this class define how a task is partitioned for resampling (e.g., in resample() or benchmark()), using a set of hyperparameters such as the number of folds in cross-validation.

Resampling objects can be instantiated on a Task, which applies the strategy on the task and manifests in a fixed partition of row_ids of the Task.

Predefined resamplings are stored in the mlr3misc::Dictionary mlr_resamplings, e.g. cv or bootstrap.

Format

R6::R6Class object.

Construction

Note: This object is typically constructed via a derived classes, e.g. ResamplingCV or ResamplingHoldout.

r = Resampling$new(id, param_set, param_vals)

id :: character(1)
Identifier for the resampling strategy.

param_set :: paradox::ParamSet
Set of hyperparameters.

param_vals :: named list()
List of hyperparameter settings.

Fields

id :: character(1)
Identifier of the learner.

param_set :: paradox::ParamSet
Description of available hyperparameters and hyperparameter settings.

hash :: character(1)
Hash (unique identifier) for this object.

instance :: any
During instantiate(), the instance is stored in this slot. The instance can be in any arbitrary format.

is_instantiated :: logical(1)
Is TRUE, if the resampling has been instantiated.

duplicated_ids :: logical(1)
Is TRUE if this resampling strategy may have duplicated row ids in a single training set or test set. E.g., this is TRUE for Bootstrap, and FALSE for cross validation.

iters :: integer(1)
Return the number of resampling iterations, depending on the values stored in the param_set.

task_hash :: character(1)
The hash of the task which was passed to r$instantiate().

Methods

instantiate(task)
Task -> self
Materializes fixed training and test splits for a given task and stores them in r$instance.

train_set(i)
integer(1) -> (integer() | character())
Returns the row ids of the i-th training set.

test_set(i)
integer(1) -> (integer() | character())
Returns the row ids of the i-th test set.

See also

Examples

r = rsmp("subsampling")

# Default parametrization
r$param_set$values

#> $repeats
#> [1] 30
#> 
#> $ratio
#> [1] 0.6666667
#> 

# Do only 3 repeats on 10% of the data
r$param_set$values = list(ratio = 0.1, repeats = 3)
r$param_set$values

#> $ratio
#> [1] 0.1
#> 
#> $repeats
#> [1] 3
#> 

# Instantiate on iris task
task = tsk("iris")
r$instantiate(task)

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>  [1]  63  56  69  60 135 125 141  13   9  51 145   3  77  86 107

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = rsmp("cv")$instantiate(task)
r$train_set(1)

#>   [1]   9  10  11  20  40  45  52  57  75  79  84  86 101 103 111   2  13  15
#>  [19]  32  47  51  60  61  82  96 130 133 134 135 146   1  24  27  33  54  81
#>  [37]  83  85  87 105 110 137 140 142 148  12  17  19  26  36  43  63  66  70
#>  [55]  90  93  97 107 108 114   6   7  23  42  46  53  67  80 109 113 118 122
#>  [73] 131 132 150  16  28  30  34  38  71  74  98 121 127 128 136 139 143 149
#>  [91]   3  18  25  29  49  58  78  89  92 104 119 123 126 144 145  14  41  44
#> [109]  48  55  56  69  88  91  95  99 106 116 120 125   5  21  22  35  37  39
#> [127]  64  73  76 100 112 115 117 129 138

# Stratification
task = tsk("pima")
prop.table(table(task$truth())) # moderately unbalanced

#> 
#>       pos       neg 
#> 0.3489583 0.6510417 

r = rsmp("subsampling")
r$instantiate(task)
prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>     pos     neg 
#> 0.34375 0.65625