Regression Task
Source: R/TaskRegr.R

This task specializes Task and TaskSupervised for regression problems. The target column is assumed to be numeric. Predefined tasks are stored in mlr_tasks.

The task_type is set to "classif".

Format

R6::R6Class object inheriting from Task/TaskSupervised.

Construction

t = TaskRegr$new(id, backend, target)

id :: character(1)
Name of the task.

backend :: DataBackend
Either a DataBackend, or any object which is convertible to a DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a DataBackendDataTable.

target :: character(1)
Name of the target column.

Fields

backend :: DataBackend.

col_info :: data.table::data.table()
Table with with 3 columns: Column names of DataBackend are stored in columnid. Column type holds the storage type of the variables, e.g. integer, numeric or character. Column levels keeps a list of possible levels for factor and character variables.

col_roles :: named list()
Each column (feature) can have an arbitrary number of roles in the learning task:

"target": Labels to predict.

"feature": Regular feature.

"order": Data returned by data() is ordered by this column (or these columns).

"groups": During resampling, observations with the same value of the variable with role "groups" are marked as "belonging together". They will be exclusively assigned to be either in the training set or the test set for each resampling iteration. Only a single column may be marked as grouping column.

"weights": Observation weights. Only a single column may be marked as weights. col_roles keeps track of the roles with a named list of vectors of feature names. To alter the roles, use t$set_col_role().

row_roles :: named list()
Each row (observation) can have an arbitrary number of roles in the learning task:

"use": Use in train / predict / resampling.

"validation": Hold the observations back unless explicitly requested. row_roles keeps track of the roles with a named list of vectors of feature names. To alter the role, use set_row_role().

feature_names :: character()
Returns all column names with role == "feature".

feature_types :: data.table::data.table()
Returns a table with columns id and type where id are the column names of "active" features of the task and type is the storage type.

formula :: formula()
Constructs a stats::formula, e.g. [target] ~ [feature_1] + [feature_2] + ... + [feature_k], using the active features of the task.

hash :: character(1)
Hash (unique identifier) for this object.

id :: character(1)
Stores the identifier of the Task.

measures :: list() of Measure
Stores the measures to use for this task.

ncol :: integer(1)
Returns the total number of cols with role "target" or "feature".

nrow :: integer(1)
Return the total number of rows with role "use".

row_ids :: (integer() | character())
Returns the row ids of the DataBackend for observations with with role "use".

target_names :: character()
Returns all column names with role "target".

task_type :: character(1)
Stores the type of the Task.

groups :: data.table::data.table()
If the task has a designated column role "groups", table with two columns: "row_id" (integer() | character()) and the grouping variable group (vector()). Returns NULL if there are is no grouping column.

weights :: data.table::data.table()
If the task has a designated column role "weights", table with two columns: "row_id" (integer() | character()) and the observation weights weight (numeric()). Returns NULL if there are is no weight column.

Methods

data(rows = NULL, cols = NULL, format = NULL)
(integer() | character(), character(), character(1)) -> any
Returns a slice of the data from the DataBackend in the format specified by format (depending on the DataBackend, but usually a data.table::data.table()). Rows are subsetted to only contain observations with role "use". Columns are filtered to only contain features with roles "target" and "feature". If invalid rows or cols are specified, an exception is raised.

cbind(data)
data.frame() -> self
Extends the DataBackend with additional columns. The row ids must be provided as column in data (with column name matching the primary key name of the DataBackend). If this column is missing, it is assumed that the rows are exactly in the order of t$row_ids.

rbind(data)
data.frame() -> self
Extends the DataBackend with additional rows. The new row ids must be provided as column in data. If this column is missing, new row ids are constructed automatically.

filter(rows)
(integer() | character()) -> self
Subsets the task, reducing it to only keep the rows specified.

select(cols)
character() -> self
Subsets the task, reducing it to only keep the columns specified.

levels(cols)
character() -> named list()
Returns the distinct levels of all columns in cols. Only applicable for features with type "character", "factor" or "ordered", and is NULL otherwise. This function ignores the row roles, it returns all levels available in the DataBackend.

head(n = 6)
integer() -> data.table::data.table()
Get the first n observations with role "use".

replace_features(data)
data.frame() -> self
Replaces some features of the task by constructing a completely new DataBackendDataTable. This operation is similar to calling select() and cbind(), but explicitly copies the data.

set_col_role(cols, new_roles, exclusive = TRUE)
(character(), character(), logical(1)) -> self
Adds the roles new_roles to columns referred to by cols. If exclusive is TRUE, the referenced columns will be removed from all other roles.

set_row_role(rows, new_roles, exclusive = TRUE)
(character(), character(), logical(1)) -> self
Adds the roles new_roles to rows referred to by rows. If exclusive is TRUE, the referenced rows will be removed from all other roles.

See also

Examples

task = TaskRegr$new("iris", backend = iris, target = "Sepal.Length")
task$task_type

#> [1] "regr"

task$formula

#> Sepal.Length ~ Petal.Length + Petal.Width + Sepal.Width + Species
#> NULL

task$truth()

#>   [1] 5.1 4.9 4.7 4.6 5.0 5.4 4.6 5.0 4.4 4.9 5.4 4.8 4.8 4.3 5.8 5.7 5.4 5.1
#>  [19] 5.7 5.1 5.4 5.1 4.6 5.1 4.8 5.0 5.0 5.2 5.2 4.7 4.8 5.4 5.2 5.5 4.9 5.0
#>  [37] 5.5 4.9 4.4 5.1 5.0 4.5 4.4 5.0 5.1 4.8 5.1 4.6 5.3 5.0 7.0 6.4 6.9 5.5
#>  [55] 6.5 5.7 6.3 4.9 6.6 5.2 5.0 5.9 6.0 6.1 5.6 6.7 5.6 5.8 6.2 5.6 5.9 6.1
#>  [73] 6.3 6.1 6.4 6.6 6.8 6.7 6.0 5.7 5.5 5.5 5.8 6.0 5.4 6.0 6.7 6.3 5.6 5.5
#>  [91] 5.5 6.1 5.8 5.0 5.6 5.7 5.7 6.2 5.1 5.7 6.3 5.8 7.1 6.3 6.5 7.6 4.9 7.3
#> [109] 6.7 7.2 6.5 6.4 6.8 5.7 5.8 6.4 6.5 7.7 7.7 6.0 6.9 5.6 7.7 6.3 6.7 7.2
#> [127] 6.2 6.1 6.4 7.2 7.4 7.9 6.4 6.3 6.1 7.7 6.3 6.4 6.0 6.9 6.7 6.9 5.8 6.8
#> [145] 6.7 6.7 6.3 6.5 6.2 5.9