Convert to a Resampling

Convert object to a Resampling or a list of Resampling.

as_resampling(x, ...)

# S3 method for Resampling
as_resampling(x, clone = FALSE, ...)

as_resamplings(x, ...)

# S3 method for list
as_resamplings(x, clone = FALSE, ...)

# S3 method for Resampling
as_resamplings(x, clone = FALSE, ...)

Arguments

x

(any)
Object to convert.

...

(any)
Additional arguments.

clone

(logical(1))
If TRUE, ensures that the returned object is not the same as the input x.