Column Auto-Converter

Set of rules to automatically convert column types. This is used during rbind-ing of Tasks, but also in some pipe operators in mlr3pipelines.

All rules are stored as functions in mlr_reflections$auto_converters.

Usage

auto_convert(value, id, type, levels)

Arguments

value

(any)
New values to convert in order to match type.

id

(character(1))
Name of the column, used in error messages.

type

(character(1))
Type to convert values to.

levels

(character() | NULL)
Levels to use for conversion to factor or ordered.

Value

Vector value converted to type type.