Generate a Benchmark Design
Source: R/benchmark.R

Takes a lists of Task, a list of Learner and a list of Resampling to generate a design in an expand.grid() fashion (a.k.a. cross join or Cartesian product).

Resampling strategies must be uninstantiated, and will be instantiated per task.

expand_grid(tasks, learners, resamplings)

Arguments

tasks
(list of Task).

learners
(list of Learner).

resamplings
(list of Resampling).

Value

(data.table()) with the cross product of the input vectors.