Generate Tasks for a Learner
Source: R/generate_tasks.R

Generates multiple tasks for a given Learner, based on its properties. This function is primarily used for unit tests, but can also assist while writing custom learners.

generate_tasks(learner, N = 20L)

Arguments

learner
(Learner).

N
(integer(1)). Number of rows of generated tasks.

Value

(list()) of Task.

Examples

tasks = generate_tasks(mlr_learners$get("classif.rpart"))
tasks$missings_binary$data()

#>     target character factor integer logical    numeric ordered
#>  1:      A         a     f1       1    TRUE 0.21516649      o1
#>  2:      B         b     f2       2      NA 0.81393372      o2
#>  3:      A         a     f1       3    TRUE 0.30776387    <NA>
#>  4:      B         b     f2       1   FALSE 0.68774269      o2
#>  5:      A         a     f1       2    TRUE 0.93268108      o1
#>  6:      B         b     f2      NA   FALSE 0.11577966      o2
#>  7:      A         a   <NA>       1    TRUE 0.12770569      o1
#>  8:      B      <NA>     f2       2   FALSE 0.67822386      o2
#>  9:      A         a     f1       3    TRUE 0.42894860      o1
#> 10:      B         b     f2       1   FALSE 0.83440098      o2
#> 11:      A         a     f1       2    TRUE 0.97143807      o1
#> 12:      B         b     f2       3   FALSE 0.07048898      o2
#> 13:      A         a     f1       1    TRUE 0.45978519      o1
#> 14:      B         b     f2       2   FALSE 0.70158512      o2
#> 15:      A         a     f1       3    TRUE 0.08694017      o1
#> 16:      B         b     f2       1   FALSE 0.99294402      o2
#> 17:      A         a     f1       2    TRUE 0.25309899      o1
#> 18:      B         b     f2       3   FALSE         NA      o2
#> 19:      A         a     f1       1    TRUE 0.68632495      o1
#> 20:      B         b     f2       2   FALSE 0.78692735      o2