Execution control object
Source: R/mlr_control.R

This function creates a named list of settings which control the execution of an Experiment. It contains all options (see mlr_options()) without the "mlr3" prefix, and additionally:

store_model: If FALSE, the model returned by the learner is discarded in order to save some memory after the experiment is completed. Note that you will be unable to further predict on new data.

store_prediction: If FALSE, the predictions are discarded in order to save some memory after the experiment is completed. Note that you will be unable calculate more performance measures.

error_handling: How to deal with models raising exceptions during train or predict?

"off" (default). An exception is raised, stopping the execution.

"catch". Exceptions are caught and logged. There will be no predictions available, and the performance will be NA. All output is stored in the Experiment as a Log.

"impute_worst". This is similar to the "catch" approach, but instead of predicting NA, the worst possible performance is predicted.

"fallback_train". If the learner fails to fit a model during train, fit a fallback model, e.g. with a featureless learner. The fallback learner is in this case used to generate predictions which are then scored. Note that this mechanism does not guard you from models which successfully train, but raise exceptions during predict. This would result in missing predictions and NA scores.

"fallback". Always fit a fallback model and use it if the learner fails to train or predict.

fallback_learner: If "error_handling" is set to "fallback_train" or "fallback", use this learner as fallback learner.

mlr_control(...)

Arguments

...
Named arguments to overwrite the defaults / options.

Examples

# get a list of the currently active defaults
mlr_control()

#> $store_model
#> [1] TRUE
#> 
#> $store_prediction
#> [1] TRUE
#> 
#> $error_handling
#> [1] "off"
#> 
#> $fallback_learner
#> NULL
#> 
#> $verbose
#> [1] TRUE
#> 
#> $debug
#> [1] FALSE
#> 
#> $use_future
#> [1] TRUE
#> 

# enable debuging
mlr_control(debug = TRUE)

#> $store_model
#> [1] TRUE
#> 
#> $store_prediction
#> [1] TRUE
#> 
#> $error_handling
#> [1] "off"
#> 
#> $fallback_learner
#> NULL
#> 
#> $verbose
#> [1] TRUE
#> 
#> $debug
#> [1] TRUE
#> 
#> $use_future
#> [1] TRUE
#>