Dictionary of Learners

A simple mlr3misc::Dictionary storing objects of class Learner. Each learner has an associated help page, see mlr_learners_[id].

This dictionary can get populated with additional learners by add-on packages. For more classification and regression learners, load the mlr3learners package.

For a more convenient way to retrieve and construct learners, see lrn().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "feature_types", "packages", "properties" and "predict_types" as columns.

See also

Examples

as.data.table(mlr_learners)

#>                    key                                    feature_types
#> 1:       classif.debug logical,integer,numeric,character,factor,ordered
#> 2: classif.featureless logical,integer,numeric,character,factor,ordered
#> 3:       classif.rpart logical,integer,numeric,character,factor,ordered
#> 4:    regr.featureless logical,integer,numeric,character,factor,ordered
#> 5:          regr.rpart logical,integer,numeric,character,factor,ordered
#>    packages                                                        properties
#> 1:                                               missings,multiclass,twoclass
#> 2:                  importance,missings,multiclass,selected_features,twoclass
#> 3:    rpart importance,missings,multiclass,selected_features,twoclass,weights
#> 4:    stats                             importance,missings,selected_features
#> 5:    rpart                     importance,missings,selected_features,weights
#>    predict_types
#> 1: response,prob
#> 2: response,prob
#> 3: response,prob
#> 4:   response,se
#> 5:      response

mlr_learners$get("classif.featureless")

#> <LearnerClassifFeatureless:classif.featureless>
#> * Model: -
#> * Parameters: method=mode
#> * Packages: -
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, character, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass

lrn("classif.rpart")

#> <LearnerClassifRpart:classif.rpart>
#> * Model: -
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, character, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights