Toy Classification Learner
Source: R/mlr_learners_classif_debug.R

A simple LearnerClassif used primarily in the unit tests and for debugging purposes. If no hyperparameter is set, it simply constantly predicts a randomly selected label. The following hyperparameters trigger the following actions:

message_train:

Outputs a message during train.

message_predict:

Outputs a message during predict.

warning_train:

Signals a warning during train.

warning_predict:

Signals a warning during predict.

error_train:

Raises an exception during train.

error_predict:

Raises an exception during predict.

segfault_train:

Provokes a segfault during train.

segfault_predict:

Provokes a segfault during predict.

x:

Numeric parameter. Ignored.

Note that segfaults may not work on your operating system. Also note that if they work, they will tear down your R session immediately!

LearnerClassifDebug

Format

R6::R6Class inheriting from LearnerClassif.