Dictionary of Performance Measures
Source: R/mlr_measures.R

A simple mlr3misc::Dictionary storing objects of class Measure. Each measure has an associated help page, see mlr_measures_[id].

This dictionary can get populated with additional measures by add-on packages.

For a more convenient way to retrieve and construct measures, see msr().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "task_type", "predict_type", and "packages" as columns.

See also

Examples

as.data.table(mlr_measures)

#>                     key task_type packages predict_type task_properties
#>  1:         classif.acc   classif  Metrics     response                
#>  2:         classif.auc   classif  Metrics         prob        twoclass
#>  3:          classif.ce   classif  Metrics     response                
#>  4:       classif.costs   classif              response                
#>  5:          classif.f1   classif  Metrics         prob        twoclass
#>  6:         classif.fdr   classif              response        twoclass
#>  7:          classif.fn   classif              response        twoclass
#>  8:         classif.fnr   classif              response        twoclass
#>  9:         classif.for   classif              response        twoclass
#> 10:          classif.fp   classif              response        twoclass
#> 11:         classif.fpr   classif              response        twoclass
#> 12:         classif.npv   classif              response        twoclass
#> 13:         classif.ppv   classif              response        twoclass
#> 14:   classif.precision   classif              response        twoclass
#> 15:      classif.recall   classif              response        twoclass
#> 16: classif.sensitivity   classif              response        twoclass
#> 17: classif.specificity   classif              response        twoclass
#> 18:          classif.tn   classif              response        twoclass
#> 19:         classif.tnr   classif              response        twoclass
#> 20:          classif.tp   classif              response        twoclass
#> 21:         classif.tpr   classif              response        twoclass
#> 22:           oob_error      <NA>              response                
#> 23:            regr.mae      regr  Metrics     response                
#> 24:            regr.mse      regr  Metrics     response                
#> 25:           regr.rmse      regr  Metrics     response                
#> 26:   selected_features      <NA>              response                
#> 27:           time_both      <NA>              response                
#> 28:        time_predict      <NA>              response                
#> 29:          time_train      <NA>              response                
#>                     key task_type packages predict_type task_properties

mlr_measures$get("classif.ce")

#> <MeasureClassifCE:classif.ce>
#> * Packages: Metrics
#> * Range: [0, 1]
#> * Minimize: TRUE
#> * Properties: -
#> * Predict type: response

msr("regr.mse")

#> <MeasureRegrMSE:regr.mse>
#> * Packages: Metrics
#> * Range: [0, Inf]
#> * Minimize: TRUE
#> * Properties: -
#> * Predict type: response