True Positive Rate

Measure to compare true observed labels with predicted labels in binary classification tasks.

Details

The True Positive Rate is defined as $$ \frac{\mathrm{TP}}{\mathrm{TP} + \mathrm{FN}}. $$ Also know as "recall" or "sensitivity".

This measure is undefined if TP + FN = 0.

Note

The score function calls mlr3measures::recall() from package mlr3measures.

If the measure is undefined for the input, NaN is returned. This can be customized by setting the field na_value.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("classif.recall")
msr("classif.recall")

Parameters

Empty ParamSet

Meta Information

Type: "binary"

Range: \([0, 1]\)

Minimize: FALSE

Required prediction: response

See also