True Negatives

Measure to compare true observed labels with predicted labels in binary classification tasks.

Details

This measure counts the true negatives, i.e. the number of predictions correctly indicating a negative class label. This is sometimes also called a "correct rejection".

Note

The score function calls mlr3measures::tn() from package mlr3measures.

If the measure is undefined for the input, NaN is returned. This can be customized by setting the field na_value.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("classif.tn")
msr("classif.tn")

Parameters

Empty ParamSet

Meta Information

Type: "binary"

Range: \([0, \infty)\)

Minimize: FALSE

Required prediction: response

See also