Debug Measure

This measure returns the number of observations in the Prediction object. Its main purpose is debugging.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("debug")
msr("debug")

Meta Information

Type: NA

Range: \([0, \infty)\)

Minimize: NA

Required prediction: 'response'

See also

Super class

mlr3::Measure -> MeasureDebug

Public fields

na_ratio

(numeric(1))
Ratio of scores which randomly should be NA, between 0 (default) and 1. Default is 0.

Methods

Public methods

MeasureDebug$new()

MeasureDebug$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureDebug$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureDebug$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = tsk("wine")
learner = lrn("classif.featureless")
measure = msr("debug")
rr = resample(task, learner, rsmp("cv", folds = 3))

#> INFO  [12:18:45.992] Applying learner 'classif.featureless' on task 'wine' (iter 1/3) 
#> INFO  [12:18:46.002] Applying learner 'classif.featureless' on task 'wine' (iter 2/3) 
#> INFO  [12:18:46.013] Applying learner 'classif.featureless' on task 'wine' (iter 3/3) 

rr$score(measure)

#>             task task_id                     learner          learner_id
#> 1: <TaskClassif>    wine <LearnerClassifFeatureless> classif.featureless
#> 2: <TaskClassif>    wine <LearnerClassifFeatureless> classif.featureless
#> 3: <TaskClassif>    wine <LearnerClassifFeatureless> classif.featureless
#>        resampling resampling_id iteration prediction debug
#> 1: <ResamplingCV>            cv         1     <list>    60
#> 2: <ResamplingCV>            cv         2     <list>    59
#> 3: <ResamplingCV>            cv         3     <list>    59