Elapsed Time Measure

Measures the elapsed time during train ("time_train"), predict ("time_predict"), or both ("time_both"). Aggregation of elapsed time defaults to mean but can be configured via the field aggregator of the Measure.

When predictions for multiple predict sets were made during resample() or benchmark(), the predict time shows the cumulative duration of all predictions. If learner$predict() is called manually, the last predict time gets overwritten. The elapsed time accounts only for the training duration of the primary learner, excluding the time required for training the fallback learner.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("time_train")
msr("time_train")

Meta Information

Task type: “NA”

Range: \([0, \infty)\)

Minimize: TRUE

Average: macro

Required Prediction: “NA”

Required Packages: mlr3

Parameters

Empty ParamSet

See also

Super class

mlr3::Measure -> MeasureElapsedTime

Public fields

stages

(character())
Which stages of the learner to measure? Usually set during construction.

Methods

Public methods

MeasureElapsedTime$new()

MeasureElapsedTime$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$obs_loss()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureElapsedTime$new(id = "elapsed_time", stages)

Arguments

id

(character(1))
Identifier for the new instance.

stages

(character())
Subset of ("train", "predict"). The runtime of provided stages will be summed.

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureElapsedTime$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.