Out-of-bag Error Measure

Returns the out-of-bag error of the Learner for learners that support it (learners with property "oob_error"). Returns NA for unsupported learners.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("oob_error")
msr("oob_error")

Meta Information

Type: NA

Range: \((-\infty, \infty)\)

Minimize: NA

Required prediction: 'response'

See also

Super class

mlr3::Measure -> MeasureOOBError

Methods

Public methods

MeasureOOBError$new()

MeasureOOBError$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureOOBError$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureOOBError$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.