Kendall's tau

Measure to compare true observed response with predicted response in regression tasks.

Details

Kendall's tau is defined as Kendall's rank correlation coefficient between truth and response. It is defined as $$ \tau = \frac{(\mathrm{number of concordant pairs)} - (\mathrm{number of discordant pairs)}}{\mathrm{(number of pairs)}} $$ Calls stats::cor() with method set to "kendall".

Note

The score function calls mlr3measures::ktau() from package mlr3measures.

If the measure is undefined for the input, NaN is returned. This can be customized by setting the field na_value.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("regr.ktau")
msr("regr.ktau")

Parameters

Empty ParamSet

Meta Information

Type: "regr"

Range: \([-1, 1]\)

Minimize: FALSE

Required prediction: response

See also