Sum of Squared Errors

Measure to compare true observed response with predicted response in regression tasks.

Details

The Sum of Squared Errors is defined as $$ \sum_{i=1}^n \left( t_i - r_i \right)^2. $$

Note

The score function calls mlr3measures::sse() from package mlr3measures.

If the measure is undefined for the input, NaN is returned. This can be customized by setting the field na_value.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("regr.sse")
msr("regr.sse")

Parameters

Empty ParamSet

Meta Information

Type: "regr"

Range: \([0, \infty)\)

Minimize: TRUE

Required prediction: response

See also