Dictionary of Resampling Strategies
Source: R/mlr_resamplings.R

A simple mlr3misc::Dictionary storing objects of class Resampling. Each resampling has an associated help page, see mlr_resamplings_[id].

This dictionary can get populated with additional resampling strategies by add-on packages.

For a more convenient way to retrieve and construct resampling strategies, see rsmp().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with columns "key", "params", and "iters".

See also

Examples

as.data.table(mlr_resamplings)

#>            key                 params iters
#> 1:   bootstrap stratify,repeats,ratio    30
#> 2:      custom                            0
#> 3:          cv         stratify,folds    10
#> 4:     holdout         stratify,ratio     1
#> 5: repeated_cv stratify,repeats,folds   100
#> 6: subsampling stratify,repeats,ratio    30

mlr_resamplings$get("cv")

#> <ResamplingCV> with 10 iterations
#> * Instantiated: FALSE
#> * Parameters: folds=10

rsmp("subsampling")

#> <ResamplingSubsampling> with 30 iterations
#> * Instantiated: FALSE
#> * Parameters: repeats=30, ratio=0.6667