Cross-Validation Resampling

Splits data using a folds-folds (default: 10 folds) cross-validation.

Dictionary

This Resampling can be instantiated via the dictionary mlr_resamplings or with the associated sugar function rsmp():

mlr_resamplings$get("cv")
rsmp("cv")

Parameters

folds (integer(1))
Number of folds.

References

Bischl B, Mersmann O, Trautmann H, Weihs C (2012). “Resampling Methods for Meta-Model Validation with Recommendations for Evolutionary Computation.” Evolutionary Computation, 20(2), 249–275. doi:10.1162/evco_a_00069 .

See also

Super class

mlr3::Resampling -> ResamplingCV

Active bindings

iters

(integer(1))
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingCV$new()

ResamplingCV$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$instantiate()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Creates a new instance of this R6 class.

Usage

ResamplingCV$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingCV$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a task with 10 observations
task = tsk("penguins")
task$filter(1:10)

# Instantiate Resampling
cv = rsmp("cv", folds = 3)
cv$instantiate(task)

# Individual sets:
cv$train_set(1)
#> [1]  1  2  7  3  5 10
cv$test_set(1)
#> [1] 4 6 8 9

# Disjunct sets:
intersect(cv$train_set(1), cv$test_set(1))
#> integer(0)

# Internal storage:
cv$instance # table
#> Key: <fold>
#>     row_id  fold
#>      <int> <int>
#>  1:      4     1
#>  2:      6     1
#>  3:      8     1
#>  4:      9     1
#>  5:      1     2
#>  6:      2     2
#>  7:      7     2
#>  8:      3     3
#>  9:      5     3
#> 10:     10     3