Wisconsin Breast Cancer Classification Task

A classification task for the mlbench::BreastCancer data set.

Column "Id" has been removed.

Column names have been converted to snake_case.

Positive class is set to "malignant".

16 incomplete cases have been removed from the data set.

All factor features have been converted to ordered factors.

Format

R6::R6Class inheriting from TaskClassif.

Construction

mlr_tasks$get("breast_cancer")
tsk("breast_cancer")

Meta Information

Task type: “classif”

Dimensions: 683x10

Properties: “twoclass”

Has Missings: FALSE

Target: “class”

Features: “bare_nuclei”, “bl_cromatin”, “cell_shape”, “cell_size”, “cl_thickness”, “epith_c_size”, “marg_adhesion”, “mitoses”, “normal_nucleoli”

See also