German Credit Classification Task
Source: R/TaskClassif_german_credit.R

A classification task for the German credit data set. The aim is to predict creditworthiness, labeled as "good" and "bad". Positive class is set to label "good".

See example for the creation of a MeasureClassifCosts as described misclassification costs.

Format

R6::R6Class inheriting from TaskClassif.

Source

Data set originally published on UCI. This is the preprocessed version taken from package evtree.

Donor: Professor Dr. Hans Hofmann
Institut für Statistik und Ökonometrie
Universität Hamburg
FB Wirtschaftswissenschaften
Von-Melle-Park 5
2000 Hamburg 13

Construction

mlr_tasks$get("german_credit")
tsk("german_credit")

See also

Examples

task = tsk("german_credit")
costs = matrix(c(0, 1, 5, 0), nrow = 2)
dimnames(costs) = list(predicted = task$class_names, truth = task$class_names)
measure = msr("classif.costs", id = "german_credit_costs", costs = costs)
print(measure)

#> <MeasureClassifCosts:german_credit_costs>
#> * Packages: -
#> * Range: [0, Inf]
#> * Minimize: TRUE
#> * Properties: requires_task
#> * Predict type: response