Palmer Penguins Data Set

Classification data to predict the species of penguins from the palmerpenguins package, see palmerpenguins::penguins. A better alternative to the iris data set.

Format

R6::R6Class inheriting from TaskClassif.

Source

palmerpenguins

Dictionary

This Task can be instantiated via the dictionary mlr_tasks or with the associated sugar function tsk():

mlr_tasks$get("penguins")
tsk("penguins")

Meta Information

Task type: “classif”

Dimensions: 344x8

Properties: “multiclass”

Has Missings: TRUE

Target: “species”

Features: “bill_depth”, “bill_length”, “body_mass”, “flipper_length”, “island”, “sex”, “year”

Pre-processing

The unit of measurement have been removed from the column names. Lengths are given in millimeters (mm), weight in gram (g).

References

Gorman KB, Williams TD, Fraser WR (2014). “Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis).” PLoS ONE, 9(3), e90081. doi: 10.1371/journal.pone.0090081 .

https://github.com/allisonhorst/palmerpenguins

See also