Manually Partition into Training and Test Set

Creates a split of the row ids of a Task into a training set and a test set while optionally stratifying on the target column.

partition(task, ratio = 0.67, stratify = TRUE, ...)

# S3 method for TaskRegr
partition(task, ratio = 0.67, stratify = TRUE, bins = 3L, ...)

# S3 method for TaskClassif
partition(task, ratio = 0.67, stratify = TRUE, ...)

Arguments

task
(Task)
Task to operate on.

ratio
(numeric(1))
Ratio of observations to put into the training set.

stratify
(logical(1))
If TRUE, stratify on the target variable. For regression tasks, the target variable is first cut into bins bins. See Task$add_strata().

...
(any)
Additional arguments, currently not used.

bins
(integer(1))
Number of bins to cut the target variable into.

Examples

# regression task
task = tsk("boston_housing")

# roughly equal size split with stratification
str(partition(task, ratio = 0.5))

#> List of 2
#>  $ train: int [1:254] 1 3 4 6 7 8 14 19 30 39 ...
#>  $ test : int [1:252] 2 13 17 38 40 41 42 43 44 45 ...

# classification task
task = tsk("pima")
split = partition(task)

# roughly same distribution of the target label
prop.table(table(task$truth(split$train)))

#> 
#>       pos       neg 
#> 0.3495146 0.6504854 

prop.table(table(task$truth(split$test)))

#> 
#>       pos       neg 
#> 0.3478261 0.6521739