Container for Results of benchmark
Source: R/BenchmarkResult.R

This is the object returned by benchmark.

Usage

bmr$tasks
bmr$learners
bmr$resamplings
bmr$measures
bmr$performance
bmr$aggregated
bmr$resample_results
bmr$resample_result(hash)

Arguments

hash [character(1)]:
String which identifies a subgroup to extract as ResampleResult.

Details

$tasks, $learners, $resamplings and $measures return an overview table of involved objects.

$performance provides a data.table() with column iteration (integer) and a numeric column for each performance measure (columns named using the measure ids).

$aggregated returns aggregated performance measures as a data.table(). The table is build similar to the one returned by $performance, but experiments are aggregated by their resample result group (combination of Task, Learner and Resampling). The actual aggregation function is defined by the respective Measure.

$resample_results returns a data.table() which gives an overview of the resample result groups in the benchmark. These groups in the BenchmarkResult can be extracted as ResampleResult for further inspection.

$resample_result() creates the ResampleResult identified by the specified hash value.

Examples

bmr = benchmark(
  tasks = mlr_tasks$mget("iris"),
  learners = mlr_learners$mget(c("classif.dummy", "classif.rpart")),
  resamplings = mlr_resamplings$mget("cv"),
  ctrl = exec_control(verbose = FALSE)
)

bmr$tasks

#>           task_hash task_id          task
#> 1: 9e2a48c8f2f07100    iris <TaskClassif>

bmr$learners

#>        learner_hash    learner_id               learner
#> 1: afef6eda1b787161 classif.rpart <LearnerClassifRpart>
#> 2: 0de784a40d0c2eea classif.dummy <LearnerClassifDummy>

bmr$resamplings

#>     resampling_hash resampling_id     resampling
#> 1: 78538621b13cb796            cv <ResamplingCV>

bmr$measures

#>    measure_id              measure
#> 1:       mmce <MeasureClassifMMCE>

bmr$performance

#>                 hash task_id    learner_id resampling_id       mmce
#>  1: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.00000000
#>  2: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.00000000
#>  3: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.13333333
#>  4: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.06666667
#>  5: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.06666667
#>  6: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.00000000
#>  7: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.00000000
#>  8: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.13333333
#>  9: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.13333333
#> 10: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.06666667
#> 11: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.73333333
#> 12: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.86666667
#> 13: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.80000000
#> 14: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.73333333
#> 15: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.66666667
#> 16: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.73333333
#> 17: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.80000000
#> 18: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.73333333
#> 19: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.80000000
#> 20: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.66666667

bmr$aggregated

#>                hash task_id    learner_id resampling_id      mmce
#> 1: 1c8a65bdbff30e7b    iris classif.rpart            cv 0.0600000
#> 2: 8ea7c1371ef8e2d6    iris classif.dummy            cv 0.7533333

rrs = bmr$resample_results
print(rrs)

#>                hash task_id    learner_id resampling_id  N
#> 1: 1c8a65bdbff30e7b    iris classif.rpart            cv 10
#> 2: 8ea7c1371ef8e2d6    iris classif.dummy            cv 10

rr = bmr$resample_result(rrs$hash[1])
print(rr)

#> ResampleResult of learner 'iris' on task 'classif.rpart' with 10 iterations
#>  Measure Min. 1st Qu.     Median Mean   3rd Qu.      Max.       Sd
#>     mmce    0       0 0.06666667 0.06 0.1166667 0.1333333 0.058373

rr$experiment(1)$model

#> n= 135 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#>  1) root 135 89 setosa (0.34074074 0.32592593 0.33333333)  
#>    2) Petal.Length< 2.45 46  0 setosa (1.00000000 0.00000000 0.00000000) *
#>    3) Petal.Length>=2.45 89 44 virginica (0.00000000 0.49438202 0.50561798)  
#>      6) Petal.Width< 1.65 47  4 versicolor (0.00000000 0.91489362 0.08510638)  
#>       12) Petal.Length< 4.85 40  0 versicolor (0.00000000 1.00000000 0.00000000) *
#>       13) Petal.Length>=4.85 7  3 virginica (0.00000000 0.42857143 0.57142857) *
#>      7) Petal.Width>=1.65 42  1 virginica (0.00000000 0.02380952 0.97619048) *