Container for Results of benchmark()
Source: R/BenchmarkResult.R

This is the result container object returned by benchmark().

Format

R6::R6Class object.

Construction

bmr = BenchmarkResult$new(data)

data :: data.table::data.table()
Table with the data of one Experiment per row. See description of field data of Experiment for the exact structure.

Fields

data :: data.table::data.table()
Experiment data with one Experiment per line.

tasks :: data.table::data.table()
Table of used tasks with three columns: "task_hash" (character(1)), "task_id" (character(1)) and "task" (Task).

learners :: data.table::data.table()
Table of used learners with three columns: "learner_hash" (character(1)), "learner_id" (character(1)) and "learner" (Learner).

resamplings :: data.table::data.table()
Table of used resamplings with three columns: "resampling_hash" (character(1)), "resampling_id" (character(1)) and "resampling" (Resampling).

measures :: data.table::data.table()
Table of used measures with three columns: "measure_hash" (character(1)), "measure_id" (character(1)) and "measure" (Measure).

resample_results :: data.table::data.table()
Table of ResampleResult groups with 5 columns: "hash" (character(1)), "task_id" (character(1)), "learner_id" (character(1)), "resamping_id" (character(1)) and "N" (integer(1)). The last column ("N") is the number of Experiments in the ResampleResult.

Methods

aggregated(objects = TRUE, ids = TRUE, params = FALSE)
(logical(1), logical(1), logical(1)) -> data.table::data.table()
Returns a result table where experiments are aggregated per ResampleResult. Arguments control the number of additional columns:

objects :: logical(1)
Return objects as columns in the result data.table().

ids :: logical(1)
Return object ids as columns in the result data.table().

params :: logical(1)
Return learner hyperparameter values as list column params in the result data.table().

combine(bmr)
BenchmarkResult -> self
Fuses a second BenchmarkResult into itself.

get_best(measure)
Measure -> ResampleResult
Returns the ResampleResult with best performance according to the provided Measure.

resample_result(hash)
character(1) -> ResampleResult
Retrieves the ResampleResult with provided hash.

S3 Methods

as.data.frame(rr)
BenchmarkResult -> data.frame()
Converts to a flat data.frame().

as.data.table(rr)
BenchmarkResult -> data.table::data.table()
Converts to a flat data.table().

Examples

set.seed(123)
bmr = benchmark(expand_grid(
  tasks = mlr_tasks$mget("iris"),
  learners = mlr_learners$mget(c("classif.featureless", "classif.rpart")),
  resamplings = mlr_resamplings$mget("cv")
))
print(bmr)

#> <BenchmarkResult> of 20 experiments in 2 resamplings:
#>  resampling task     learner classif.mmce
#>          cv iris       rpart   0.07333333
#>          cv iris featureless   0.80666667
#> 
#> Public: aggregated(), clone(), combine(), data, get_best(), learners,
#>   measures, resample_result(), resample_results, resamplings, tasks

bmr$tasks

#>           task_hash task_id          task
#> 1: edef5f3529818e77    iris <TaskClassif>

bmr$learners

#>        learner_hash  learner_id                     learner
#> 1: a3aea4ed7bc571d2 featureless <LearnerClassifFeatureless>
#> 2: d64806633203a6d5       rpart       <LearnerClassifRpart>

bmr$resamplings

#>     resampling_hash resampling_id     resampling
#> 1: e740f22253d1c2ee            cv <ResamplingCV>

bmr$measures

#>        measure_hash   measure_id              measure
#> 1: c7a66ec863ba061f classif.mmce <MeasureClassifMMCE>

# aggregated results
bmr$aggregated(objects = FALSE)

#>                hash resampling_id task_id  learner_id classif.mmce
#> 1: 50d4d40aa929f6b1            cv    iris featureless   0.80666667
#> 2: e23c7afd2a3e4f9c            cv    iris       rpart   0.07333333

# aggregated results with hyperparameters as separate columns
mlr3misc::unnest(bmr$aggregated(objects = FALSE, params = TRUE), "params")

#>                hash resampling_id task_id  learner_id classif.mmce method
#> 1: 50d4d40aa929f6b1            cv    iris featureless   0.80666667   mode
#> 2: e23c7afd2a3e4f9c            cv    iris       rpart   0.07333333   <NA>

# extract resample results and experiments
rrs = bmr$resample_results
print(rrs)

#>                hash task_id  learner_id resampling_id  N
#> 1: 50d4d40aa929f6b1    iris featureless            cv 10
#> 2: e23c7afd2a3e4f9c    iris       rpart            cv 10

rr = bmr$resample_result(rrs$hash[1])
print(rr)

#> <ResampleResult> of learner 'iris' on task 'featureless' with 10 iterations
#>       Measure   Min. 1st Qu. Median   Mean 3rd Qu.   Max.      Sd
#>  classif.mmce 0.7333     0.8    0.8 0.8067     0.8 0.9333 0.05837

rr$experiment(1)$model

#> $tab
#> 
#>     setosa versicolor  virginica 
#>         49         44         42 
#> 
#> $features
#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> attr(,"class")
#> [1] "featureless"