Container for Results of benchmark()
Source: R/BenchmarkResult.R

This is the result container object returned by benchmark().

Format

R6Class object.

Usage

# Construction
# Users of mlr3 generally don't need to construct this class themselves, but rather use objects returned by [benchmark()].
bmr = benchmark(...)
    # Members
bmr$aggregated
bmr$data
bmr$learners
bmr$measures
bmr$resample_results
bmr$resamplings
bmr$tasks
    # Methods
bmr$combine(bmr)
bmr$get_best(measure)
bmr$resample_result(hash)
    # S3 methods
as.data.table(bmr)

Arguments

data (data.table()): Data to create the BenchmarkResult. Each line must contain data of a single Experiment (complex objects wrapped inside lists to create list columns).

hash (character(1)): String which identifies a subgroup to extract as ResampleResult.

bmr (BenchmarkResult).

measure (Measure).

Details

$aggregated returns aggregated performance measures as a data.table(). Experiments are aggregated by their resample result group (combination of Task, Learner and Resampling). The actual aggregation function is defined by the respective Measure.

$combine() takes a second BenchmarkResult bmr as argument and extends itself with its data.

$data returns the full benchmark structure for each iteration (task, learner, resampling, etc).

$resample_results returns a data.table() which gives an overview of the resample result groups in the benchmark. These groups in the BenchmarkResult can be extracted as ResampleResult for further inspection.

$tasks, $learners, $resamplings and $measures return an overview table of included objects, together with a unique hash for the respective object.

as.data.table() converts a BenchmarkResult to a data.table().

$resample_result() creates the ResampleResult identified by the specified hash value.

Examples

bmr = benchmark(expand_grid(
  tasks = mlr_tasks$mget("iris"),
  learners = mlr_learners$mget(c("classif.featureless", "classif.rpart")),
  resamplings = mlr_resamplings$mget("cv")
))
print(bmr)

#> <BenchmarkResult> of 20 experiments in 2 resamplings:
#>  task     learner resampling classif.mmce
#>  iris       rpart         cv   0.07333333
#>  iris featureless         cv   0.80666667
#> 
#> Public: aggregated, clone(), combine(), data, get_best(), learners,
#>   measures, resample_result(), resample_results, resamplings, tasks

bmr$tasks

#>           task_hash task_id          task
#> 1: 42ae494eae273e58    iris <TaskClassif>

bmr$learners

#>        learner_hash  learner_id                     learner
#> 1: e36b506ae4431e12 featureless <LearnerClassifFeatureless>
#> 2: d6035d6cb5b51eed       rpart       <LearnerClassifRpart>

bmr$resamplings

#>     resampling_hash resampling_id     resampling
#> 1: 4ffbfa1366d8d553            cv <ResamplingCV>

bmr$measures

#>      measure_id              measure
#> 1: classif.mmce <MeasureClassifMMCE>

bmr$aggregated

#>                hash  resample_result task_id  learner_id resampling_id
#> 1: c3713f1a2fb079b2 <ResampleResult>    iris featureless            cv
#> 2: 36c27636bc347c19 <ResampleResult>    iris       rpart            cv
#>    classif.mmce
#> 1:   0.80666667
#> 2:   0.07333333

rrs = bmr$resample_results
print(rrs)

#>                hash task_id  learner_id resampling_id  N
#> 1: c3713f1a2fb079b2    iris featureless            cv 10
#> 2: 36c27636bc347c19    iris       rpart            cv 10

rr = bmr$resample_result(rrs$hash[1])
print(rr)

#> <ResampleResult> of learner 'iris' on task 'featureless' with 10 iterations
#>       Measure   Min. 1st Qu. Median   Mean 3rd Qu.   Max.      Sd
#>  classif.mmce 0.7333     0.8    0.8 0.8067     0.8 0.9333 0.05837

rr$experiment(1)$model

#> $tab
#> 
#>     setosa versicolor  virginica 
#>         49         44         42 
#> 
#> $features
#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width" 
#> 
#> attr(,"class")
#> [1] "featureless"