DataBackend
Source: R/DataBackend.R

This is the abstract base class for data backends. All objects of type DataBackend must provide the following interface:

Usage

b = as_data_backend(data)
b$data(rows, cols)
b$head(n = 6)
b$distinct(cols)
b$missing(rows, cols)
b$rownames
b$colnames
b$nrow
b$ncol
print(b)

See DataBackendDataTable for an exemplary implementation of this interface.

Arguments

data:
Data to wrap the backend around. Typically a data.frame.

rows (integer() or character()):
Vector of row indices to subset rows using the primary key in the data backend.

cols (character()):
Vector of column names to select.

n (integer(1)):
Number of rows to return.

Details

as_data_backend() wraps a DataBackend around the provided data. See methods("as_data_backend") for possible input formats.

$data() (data.table::data.table()) returns a slice of the data: rows are filtered using the primary_key column, columns are selected by name.

$head() (data.table::data.table()) returns a table of the first n data rows.

$distinct() (named list()) returns distinct values for columns cols.

$missing() (named integer()) returns a named integer with the number of missing values per column.

$rownames (integer() | character()) returns all row names of data.

$colnames (character()) returns all column names of data.

$nrow (integer(1)) returns the number of total rows.

$ncol (integer(1)) returns the number of total columns, including primary key column.

See also