DataBackend for Matrix
Source: R/DataBackendMatrix.R

Abstraction for Matrix(). Data is stored as (sparse) matrix. Supports two output formats:

"data.table" (default): Returns a data.table::data.table(). The primary key is returned as a regular column.

"sparse" (native): Returns a Matrix::Matrix(). The primary key is stored as additional attribute ..row_id.

Usage

# Construction
b = DataBackendMatrix$new(data)
b = as_data_backend(data)

The interface is described in DataBackend.

Arguments

data (Matrix::Matrix()).
A (sparse) matrix. If data has row names, these will be used as primary key. Integer keys (seq_len(nrow(data))) are used otherwise.

See also

Examples

requireNamespace("Matrix")

#> Loading required namespace: Matrix

data = Matrix::Matrix(sample(0:1, 20, replace = TRUE), ncol = 2)
colnames(data) = c("x1", "x2")
rownames(data) = paste0("row_", 1:10)

b = as_data_backend(data)
print(b$head(n = 3, format = "data.table"))

#>    x1 x2 ..row_id
#> 1:  1  1    row_1
#> 2:  0  0    row_2
#> 3:  0  1    row_3

print(b$head(n = 3, format = "sparse"))

#> 3 x 2 sparse Matrix of class "dgCMatrix"
#>       x1 x2
#> row_1  1  1
#> row_2  .  .
#> row_3  .  1