Key-Value Storage
Source: R/Dictionary.R

A simple key-value store for R6 objects. On retrieval of an object, the following applies:

R6 Factories (objects of class R6ClassGenerator) are initialized (with no arguments).

R6::R6Class objects are cloned.

Functions are called (with no arguments).

All other objects are returned as-is.

Usage

# Construction
d = Dictionary$new()
# Getters
d$keys(pattern)
d$add(value)
d$get(key)
d$mget(keys)
d$has(key)
d$remove(key, value)
d$remove(keys)
# S3 methods
as.data.frame(d)
as.data.table(d)
names(d)
length(d)
d[[key]]
d[keys]

Arguments

pattern (character(1)):
Restrict keys to keys which match pattern.

key (character(1)):
Single Key as string.

value:
Arbitrary value.

keys (character()):
Vector of keys.

Details

$new() initializes a new object of class Dictionary.

$keys() (character()) returns a vector with all keys (or all keys matching pattern).

$get() retrieves a single object with key key (or raises an exception).

$mget() (named list) creates a list of objects with keys keys (or raises an exception).

$has() (logical()) is TRUE if key is present in the Dictionary.

$add() adds item value with key key to the Dictionary.

$remove() removes item with key key from the Dictionary.

as.data.frame() and as.data.table() give a summarizing overview as data.frame or data.table, respectively.

names(d) is an alternative way to call d$keys().

d[[key]] is an alternative way to call d$get(key).

d[keys] is an alternative way to call d$mget(key).

length(d) returns the number of items stored in the Dictionary.

See also