Experiment
Source: R/Experiment.R

Container object for machine learning experiments.

Usage

# Construction
e = Experiment$new(task, learner, ...)
# Stepwise execution
e$train(subset)
e$predict(subset, newdata)
e$score(measures = NULL)
# Getters
e$ctrl
e$task
e$learner
e$model
e$prediction
e$performance
e$train_set
e$test_set
e$validation_set
e$timings
e$logs
e$state
e$data
e$hash
e$has_errors

Arguments

task (Task):
Task to conduct experiment on

learner (Learner):
Learner to conduct experiment with.

subset (integer() | character()):
Subset of the task's row ids to work on.

newdata (data.frame()):
New data to predict on. Will be appended to the task.

measures (list of Measure):
Performance measure to use. Defaults to the measures set in the Task.

Details

$new() initializes a new machine learning experiment which can grow in a stepwise fashion.

$task and $learner can be used to access the Task and Learner.

$train() fits the induced Learner on the (subset of the) task and stores the model in the Learner. The model can be accessed via e$model.

$predict() uses the previously fitted model to predict new observations. The predictions are stored internally as an Prediction object and can be accessed via e$prediction as data.table::data.table().

$score() quantifies stored predictions using the provided list of Measure (or the task's Measure if not provided) and stores the resulting performance values. The performance can be accessed via e$performance.

$train_set and $test_set (integer() | character()) return the row ids of the training set or test set, respectively.

$validation_set (integer() | character()) returns the row ids of the validation set (see Task).

$timings (named numeric(3)) holds the elapsed time for the steps train, predict and score in seconds with up to millisecond accuracy (c.f. proc.time()). Timings are NA if the respective step has not been performed yet.

$logs (named list(2)) returns a list with names train and predict. Both store an object of class Log if logging of the learner has been enabled via mlr_control(), and are NULL if logging was disabled or the respective step has not been performed yet.

$state (ordered(1)) returns the state of the experiment: "defined", "trained", "predicted", or "scored".

$data stores the internal representation of an Experiment as a named list with the following slots:

task (Task).

learner (Learner).

resampling (Resampling). Is NULL prior to calling $train(). If the experiment is constructed manually (i.e., not via resample() or benchmark(), a ResamplingCustom object is stored here.

iteration (integer(1)). If the experiment is constructed manually, this is always 1.

train_log: Log for the training step.

train_time (numeric(1)). Elapsed time in microseconds.

predict_log. Log for the predict step.

predict_time (numeric(1)). Elapsed time in microseconds.

prediction (Prediction).

measures (list of Measure). Actually used performance measures.

performance (named numeric). Performance values are returned by the measures.

score_time (numeric(1)). Elapsed time in microseconds.

FIXME: Missing: has_errors, hash, ctrl

Examples

e = Experiment$new(
  task = mlr_tasks$get("iris"),
  learner = mlr_learners$get("classif.rpart")
)
print(e)

#> <Experiment> [defined]:
#>  + Task: iris
#>  + Learner: classif.rpart
#>  - Model: [missing]
#>  - Predictions: [missing]
#>  - Performance: [missing]
#> 
#> Public: clone(), ctrl, data, has_errors, hash, learner, logs, model,
#>   performance, predict(), prediction, score(), state, task, test_set,
#>   timings, train_set, train(), validation_set

e$state

#> [1] defined
#> Levels: defined < trained < predicted < scored

e$train(subset = 1:120)

#> INFO [mlr3] Training learner 'classif.rpart' on task 'iris' ...

print(e)

#> <Experiment> [trained]:
#>  + Task: iris
#>  + Learner: classif.rpart
#>  + Model: [rpart]
#>  - Predictions: [missing]
#>  - Performance: [missing]
#> 
#> Public: clone(), ctrl, data, has_errors, hash, learner, logs, model,
#>   performance, predict(), prediction, score(), state, task, test_set,
#>   timings, train_set, train(), validation_set

e$state

#> [1] trained
#> Levels: defined < trained < predicted < scored

e$model

#> n= 120 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 120 70 setosa (0.41666667 0.41666667 0.16666667)  
#>   2) Petal.Length< 2.45 50  0 setosa (1.00000000 0.00000000 0.00000000) *
#>   3) Petal.Length>=2.45 70 20 versicolor (0.00000000 0.71428571 0.28571429)  
#>     6) Petal.Length< 4.95 49  1 versicolor (0.00000000 0.97959184 0.02040816) *
#>     7) Petal.Length>=4.95 21  2 virginica (0.00000000 0.09523810 0.90476190) *

e$predict(subset = 121:150)

#> INFO [mlr3] Predicting with model of learner 'classif.rpart' on task 'iris' ...

print(e)

#> <Experiment> [predicted]:
#>  + Task: iris
#>  + Learner: classif.rpart
#>  + Model: [rpart]
#>  + Predictions: [PredictionClassif]
#>  - Performance: [missing]
#> 
#> Public: clone(), ctrl, data, has_errors, hash, learner, logs, model,
#>   performance, predict(), prediction, score(), state, task, test_set,
#>   timings, train_set, train(), validation_set

e$state

#> [1] predicted
#> Levels: defined < trained < predicted < scored

e$prediction

#> <PredictionClassif> for 30 observations:
#>     row_id   response     truth
#>  1:    121  virginica virginica
#>  2:    122 versicolor virginica
#>  3:    123  virginica virginica
#> ---                            
#> 28:    148  virginica virginica
#> 29:    149  virginica virginica
#> 30:    150  virginica virginica

e$score()

#> INFO [mlr3] Scoring predictions of learner 'classif.rpart' on task 'iris' ...

print(e)

#> <Experiment> [scored]:
#>  + Task: iris
#>  + Learner: classif.rpart
#>  + Model: [rpart]
#>  + Predictions: [PredictionClassif]
#>  + Performance: mmce=0.166667
#> 
#> Public: clone(), ctrl, data, has_errors, hash, learner, logs, model,
#>   performance, predict(), prediction, score(), state, task, test_set,
#>   timings, train_set, train(), validation_set

e$state

#> [1] scored
#> Levels: defined < trained < predicted < scored

e$performance

#>      mmce 
#> 0.1666667 

e$train_set

#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120

e$test_set

#>  [1] 121 122 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139
#> [20] 140 141 142 143 144 145 146 147 148 149 150