Regression Learner

This Learner specializes Learner for regression problems.

Many predefined learners can be found in the mlr3misc::Dictionary mlr_learners after loading the mlr3learners package.

Format

R6::R6Class object inheriting from Learner.

Construction

l = LearnerRegr$new(id, param_set = ParamSet$new(), predict_types = character(), feature_types = character(),
    properties = character(), data_formats = "data.table", packages = character(), man = NA_character_)

For a description of the arguments, see Learner. task_type is set to "regr".

Possible values for predict_types are passed to and converted by PredictionRegr:

"response": Predicts a numeric response for each observation in the test set.

"se": Predicts the standard error for each value of response for each observation in the test set.

Fields

See Learner.

Methods

See Learner.

See also

Examples

# get all regression learners from mlr_learners:
lrns = mlr_learners$mget(mlr_learners$keys("^regr"))
names(lrns)

#> [1] "regr.featureless" "regr.rpart"      

# get a specific learner from mlr_learners:
mlr_learners$get("regr.rpart")

#> <LearnerRegrRpart:regr.rpart>
#> * Model: -
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, selected_features, weights

lrn("classif.featureless")

#> <LearnerClassifFeatureless:classif.featureless>
#> * Model: -
#> * Parameters: method=mode
#> * Packages: -
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, character, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass