Learner Output Log

Object which stores the output of the train or predict step of an Experiment.

Log

Format

An object of class R6ClassGenerator of length 24.

Usage

l = Log$new(log = NULL)
l$is_empty
l$has_condition(cl)
    l$format()
l$print()

Arguments

log (data.table()): Object as returned by evaluate::evaluate().

cl (character(1)): Class of a condition. One of "output", "message", "warning", or "error".

Details

$new(log) parses the object returned by evaluate::evaluate() and creates a new Log.

$is_empty is TRUE if there has been no log captured, either because logging was turned off or there was not a single line out output.

$has_condition(cl) returns TRUE if at least on message of class cl is logged. Possible conditions are "output", "message", "warning", and "error".

format() and print() are for formatting and printing via base::format() or base::print(), respectively.

Examples

# Create a simple experiment and extract the train log:
e = Experiment$new(
 task = mlr_tasks$get("sonar"),
 learner = mlr_learners$get("classif.crashtest")
)
e$train(ctrl = exec_control(use_evaluate = TRUE))

#> Training learner 'classif.crashtest' on task 'sonar' ...

log = e$data$train_log

log$is_empty

#> [1] FALSE

log$has_condition("error")

#> [1] TRUE

log$print()

#> <Log> with 1 message:
#> 1: [error] Error in classif.crashtest during train()