Learner Output Log
Source: R/Log.R

Object which stores the output of the train or predict step of an Experiment.

Usage

# Construction
l = Log$new(log = NULL)
#
l$has_condition(cl)
l$warnings
l$errors
l$format()
l$print()

Arguments

log (data.table::data.table()): data.table() with columns class (character()) and message (character()).

cl (character(1)): Class of a condition. One of "output", "warning", or "error".

Details

$new(log) parses the object returned by evaluate::evaluate() and creates a new Log.

$has_condition(cl) returns TRUE if at least on message of class cl is logged. Possible conditions are "output", "message", "warning", and "error".

$warnings (character()) returns all lines which are warnings.

$errors (character()) returns all lines which are errors.

format() and print() are for formatting and printing via format() or print(), respectively.

Examples

# Create a simple experiment and extract the train log:
task = mlr_tasks$get("sonar")
learner = mlr_learners$get("classif.debug")
learner$param_vals = list(message_train = TRUE, error_train = TRUE)
learner$fallback = mlr_learners$get("classif.featureless")
e = Experiment$new(task, learner)
e$train(ctrl = list(encapsulate_train = "evaluate"))

#> INFO [mlr3] Training learner 'classif.debug' on task 'sonar' ...

l = e$logs$train

l$has_condition("error")

#> [1] TRUE

print(l)

#> <Log> with 2 messages:
#> 1: [output] Message from classif.debug->train()
#> 2: [error] Error from classif.debug->train()