Abstract measure class

Abstraction for performance measures.

Predefined measures are stored in mlr_measures.

Usage

m = Measure$new(id)
m$id
m$packages
m$task_type
m$task_properties
m$learner_properties
m$calculate(e)
m$range
m$minimize
m$aggregate(rr)

Arguments

id (character(1)): identifier for this object.

e (Experiment): Experiment to work on.

rr (ResampleResult): Performance object returned by resample to be aggregated.

Details

$new() creates a new object.

$id (character(1)) stores the identifier of the object.

$packages (character(1)) stores the names of required packages.

$task_type (character) stores the class names of tasks this measure can operate on.

$task_properties (character) defines a set of required task properties.

$learner_properties (character) defines a set of required learner properties.

$range (numeric(2)) stores the numeric range of feasible measure values.

$minimize (logical(1)) indicates if the good values are reached via minimization.

$calculate (function) does the actual work.

$aggregate (function) aggregates multiple performance measures using the aggregate function. Operates on a ResampleResult as returned by resample.

See also

Examples

mlr_measures$get("mmce")

#> <MeasureClassifMMCE>
#>   Inherits from: <MeasureClassif>
#>   Public:
#>     aggregate: function (rr) 
#>     calculate: function (e) 
#>     clone: function (deep = FALSE) 
#>     id: mmce
#>     initialize: function (id = "mmce") 
#>     learner_properties: 
#>     minimize: TRUE
#>     packages: Metrics
#>     predict_type: response
#>     range: 0 1
#>     task_properties: 
#>     task_type: classif