Abstract Prediction Object

This is the abstract base class for task objects like PredictionClassif or PredictionRegr.

Prediction objects store the following information:

The row ids of the test set

The corresponding true (observed) response.

The corresponding predicted response.

Additional predictions based on the class and predict_type. E.g., the class probabilities for classification or the estimated standard error for regression.

Note that this object is usually constructed via a derived classes, e.g. PredictionClassif or PredictionRegr.

S3 Methods

as.data.table(rr)
Prediction -> data.table::data.table()
Converts the data to a data.table::data.table().

c(..., keep_duplicates = TRUE)
(Prediction, Prediction, ...) -> Prediction
Combines multiple Predictions to a single Prediction. If keep_duplicates is FALSE and there are duplicated row ids, the data of the former passed objects get overwritten by the data of the later passed objects.

See also

Public fields

data

(named list())
Internal data structure.

task_type

(character(1))
Required type of the Task.

task_properties

(character())
Required properties of the Task.

predict_types

(character())
Set of predict types this object stores.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. Defaults to NA, but can be set by child classes.

Active bindings

row_ids

(integer())
Vector of row ids for which predictions are stored.

truth

(any)
True (observed) outcome.

missing

(integer())
Returns row_ids for which the predictions are missing or incomplete.

Methods

Public methods

Prediction$format()

Prediction$print()

Prediction$help()

Prediction$score()

Prediction$filter()

Prediction$clone()

Method format()

Helper for print outputs.

Usage

Prediction$format()

Method print()

Printer.

Usage

Prediction$print(...)

Arguments

...

(ignored).

Method help()

Opens the corresponding help page referenced by field $man.

Usage

Prediction$help()

Method score()

Calculates the performance for all provided measures Task and Learner may be NULL for most measures, but some measures need to extract information from these objects. Note that the predict_sets of the measures are ignored by this method, instead all predictions are used.

Usage

Prediction$score(
  measures = NULL,
  task = NULL,
  learner = NULL,
  train_set = NULL
)

Arguments

measures

(Measure | list of Measure)
Measure(s) to calculate.

task

(Task).

learner

(Learner).

train_set

(integer()).

Returns

Prediction.

Method filter()

Filters the Prediction, keeping only predictions for the provided row_ids. This changes the object in-place, you need to create a clone to preserve the original Prediction.

Usage

Prediction$filter(row_ids)

Arguments

row_ids

integer()
Row indices.

Returns

self, modified.

Method clone()

The objects of this class are cloneable with this method.

Usage

Prediction$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.