Prediction Object for Classification
Source: R/PredictionClassif.R

This object stores the predictions returned by a learner of class LearnerClassif. If probabilities are provided via construction and response is missing, the response is calculated from the probabilities: the class label with the highest probability is chosen. In case of ties, a random class label of the tied labels picked.

Format

R6::R6Class object

Usage

Inherits from Prediction

# Construction
p = PredictionClassif$new(task, response, prob = NULL)
    # Members
p$predict_types
p$response
p$row_ids
p$truth
    # S3 methods
as.data.table(p)

Arguments

task (Task):
Task used for prediction. Used to extract row_ids and truth. Set to NULL to skip all argument checks during initialization. Slots p$row_ids and p$truth need to be set manually in this case

response (factor() | character()):
Vector of predicted class labels.

prob (matrix):
Numeric matrix of class probabilities with one column for each class in task$all_classes and one row for each observation in the test set.

Details

$predict_types (character) stores the predict types available: a subset of c("response", "se").

$response stores the predicted values.

row_ids stores the row IDs.

$truth stores the true values.

$new() initializes a new object of class Prediction.

The prediction object can be transformed to a simple data.table::data.table with data.table::as.data.table.

References

HTML help page

See also

Examples

task = mlr_tasks$get("iris")
learner = mlr_learners$get("classif.rpart")
learner$predict_type = "prob"
e = Experiment$new(task, learner)$train()$predict()

#> INFO [mlr3] Training learner 'classif.rpart' on task 'iris' ...
#> INFO [mlr3] Predicting with model of learner 'classif.rpart' on task 'iris' ...

p = e$prediction
p$predict_types

#> [1] "response" "prob"    

head(as.data.table(p))

#>    row_id response  truth prob.setosa prob.versicolor prob.virginica
#> 1:      1   setosa setosa           1               0              0
#> 2:      2   setosa setosa           1               0              0
#> 3:      3   setosa setosa           1               0              0
#> 4:      4   setosa setosa           1               0              0
#> 5:      5   setosa setosa           1               0              0
#> 6:      6   setosa setosa           1               0              0