Prediction Object for Regression
Source: R/PredictionRegr.R

This object wraps the predictions returned by a learner of class LearnerRegr, i.e. the predicted response and standard error.

Format

R6::R6Class object inheriting from Prediction.

Construction

p = PredictionRegr$new(task = NULL, row_ids = task$row_ids, truth = task$truth(), response = NULL, se = NULL)

task :: TaskRegr
Task, used to extract defaults for row_ids and truth.

row_ids :: (integer() | character())
Row ids of the observations in the test set.

truth :: numeric()
True (observed) response.

response :: numeric()
Vector of numeric response values. One element for each observation in the test set.

se :: numeric()
Numeric vector of predicted standard error. One element for each observation in the test set.

Fields

All fields from Prediction, and additionally:

response :: numeric()
Access to the stored predicted response.

se :: numeric()
Access to the stored standard error.

The field task_type is set to "regr".

See also

Examples

task = tsk("boston_housing")
learner = lrn("regr.featureless", predict_type = "se")
p = learner$train(task)$predict(task)
p$predict_types

#> [1] "response" "se"      

head(as.data.table(p))

#>    row_id truth response       se
#> 1:      1  24.0 22.53281 9.197104
#> 2:      2  21.6 22.53281 9.197104
#> 3:      3  34.7 22.53281 9.197104
#> 4:      4  33.4 22.53281 9.197104
#> 5:      5  36.2 22.53281 9.197104
#> 6:      6  28.7 22.53281 9.197104