Prediction Object for Regression
Source: R/PredictionRegr.R

This object stores the predictions returned by a learner of class LearnerRegr.

Format

R6::R6Class object

Usage

Inherits from Prediction

# Construction
p = PredictionRegr$new(task = NULL, response = NULL, se = NULL)
    # Members
p$predict_types
p$response
p$row_ids
p$se
p$truth
    # S3 methods
as.data.table(p)

Arguments

task (Task):
Task used for prediction. Used to extract row_ids and truth. Set to NULL to skip all argument checks during initialization. Slots p$row_ids and p$truth need to be set manually in this case

truth (numeric()):
Numeric vector of true response.

response (numeric()):
Numeric vector of predictions. One element for each observation in the test set.

se (numeric()):
Numeric vector of predicted standard error. One element for each observation in the test set.

Details

$predict_types (character) stores the predict types available: a subset of c("response", "se").

$response stores the predicted values.

row_ids stores the row IDs.

$se stores the predicted standard errors (if available), or is NULL.

$truth stores the true values.

$new() initializes a new object of class Prediction.

The prediction object can be transformed to a simple data.table::data.table with data.table::as.data.table.

References

HTML help page

See also

Examples

task = mlr_tasks$get("bh")
learner = mlr_learners$get("regr.featureless")
learner$predict_type = "se"
e = Experiment$new(task, learner)$train()$predict()

#> INFO [mlr3] Training learner 'regr.featureless' on task 'boston_housing' ...
#> INFO [mlr3] Predicting with model of learner 'regr.featureless' on task 'boston_housing' ...

p = e$prediction
p$predict_types

#> [1] "response" "se"      

head(as.data.table(p))

#>    row_id response truth       se
#> 1:      1 22.53281  24.0 9.197104
#> 2:      2 22.53281  21.6 9.197104
#> 3:      3 22.53281  34.7 9.197104
#> 4:      4 22.53281  33.4 9.197104
#> 5:      5 22.53281  36.2 9.197104
#> 6:      6 22.53281  28.7 9.197104