Container for Results of resample

This is the object returned by resample.

Usage

rr = ResampleResult$new(data, hash = NULL)
    rr$task
rr$learner
rr$resampling
rr$measures
rr$experiment(iter)
rr$experiments(iters)
rr$combine(rr)
rr$performance
rr$aggregated
rr$hash

Arguments

data (data.table): data.table with columns matching the data of an Experiment. Each row corresponds to a single experiment.

hash (NULL or character(1)): Pre-calculated hash for the combination of task, learner and resampling. If NULL, the checksum will be calculated on-demand.

i (integer): Iteration(s) of the experiment(s) to retrieve.

rr (ResampleResult): Second ResampleResult.

Details

$task, learner, resampling and measure allow access to the Task, Learner, Resampling and Measure used in the resampling.

$experiment() returns an Experiment for the iter-th resampling iteration.

$experiments() returns a list with the slice of Experiments for the provided iters.

$combine() takes a second ResampleResult and combines both ResampleResults into a BenchmarkResult.

$performance provides a data.table::data.table with column iteration (integer) and a numeric column for each performance measure (columns named using the measure ids).

$aggregated returns the aggregated performance measures. The aggregation method is part of the Measure.

$hash stores a hash for the combination of task, learner and resampling.