Resampling Class
Source: R/Resampling.R

Abstraction for resampling strategies. Predefined resamplings are stored in mlr_resamplings.

Format

R6Class object.

Usage

# Construction
r = Resampling$new(id, param_set, param_vals)
    # Members
r$duplicated_ids
r$hash
r$id
r$instance
r$is_instantiated
r$iters
r$param_set
r$stratify
r$task_hash
    # Methods
r$instantiate(task)
r$test_set(i)
r$train_set(i)

Arguments

id (character(1)): identifier for this object.

param_set (paradox::ParamSet): describes available parameters.

param_vals (list()): Named list of parameter values.

task (Task): Object of type Task.

i (integer(1)): Get the i-th training/test set.

Details

$duplicated_ids is TRUE if the resampling allows observations to be included multiple times in the training set. E.g., this is true for bootstrapping, but not for cross validation.

$hash (character(1)) stores a checksum calculated on the id, param_vals and the instantiation. If the object is not instantiated yet, NA is returned.

$id (character(1)) stores the identifier of the object.

$instance stores the instantiated realization of the resampling. This is an arbitrary object, do not work directly with it. Instead, use $train_set() and $test_set().

$instantiate materializes fixed training and test splits for a given task.

$is_instantiated returns TRUE if the resampling has been instantiated, and FALSE otherwise.

$iters (integer(1)) calculates the resulting number of iterations, given the current param_vals.

$new() creates a new object of class Resampling.

$param_set (paradox::ParamSet) describes available parameters.

$stratify can be set to column names of the Task which will be used for stratification during instantiation.

$task_hash stores the hash of the task for which the resampling has been instantiated.

$test_set() returns the test set for the i-th iteration.

$train_set() returns the training set for the i-th iteration.

See also

Examples

r = mlr_resamplings$get("subsampling")

# Default parametrization
r$param_set$param_vals

#> NULL

# Do only 3 repeats on 10% of the data
r$param_set$param_vals = list(ratio = 0.1, repeats = 3)

#> Error in r$param_set$param_vals = list(ratio = 0.1, repeats = 3): cannot add bindings to a locked environment

r$param_set$param_vals

#> NULL

# Instantiate on iris task
task = mlr_tasks$get("iris")
r$instantiate(task)

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>   [1]   1   2   4   5   6   7   8  10  11  13  15  16  17  19  20  21  22  23
#>  [19]  25  26  28  29  31  32  34  36  37  38  42  43  44  45  46  52  53  56
#>  [37]  57  60  61  63  64  66  67  68  69  72  74  75  76  77  81  82  84  86
#>  [55]  87  88  89  91  92  93  94  95  96 100 103 104 105 107 108 109 110 111
#>  [73] 112 113 114 115 117 118 119 120 122 124 127 128 130 131 133 134 135 137
#>  [91] 138 139 140 141 143 144 145 146 148 149

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = mlr_resamplings$get("cv")$instantiate(task)
r$train_set(1)

#>   [1]   1  20  31  55  58  59  70  86 107 116 123 124 133 137 147   7  30  36
#>  [19]  38  51  52  60  65  84  99 104 111 113 127 138   9  23  28  37  49  53
#>  [37]  54  76  77  82  85  87 100 103 136   2  10  11  12  18  24  34  47  66
#>  [55]  75  78  88 117 131 135  13  40  61  71  80  83  92  93 110 112 119 129
#>  [73] 130 143 144   5  16  21  22  26  32  35  41  64  69  74  79  81 102 121
#>  [91]   3  15  17  25  48  50  73  91  94 101 106 118 145 149 150   8  14  29
#> [109]  33  39  42  46  62  72  98 105 122 125 126 142   6  19  27  43  45  57
#> [127]  68  89  90  95 108 120 128 134 148

# Stratification
task = mlr_tasks$get("pima")
prop.table(table(task$truth())) # moderately unbalanced

#> 
#>       neg       pos 
#> 0.6510417 0.3489583 

r = mlr_resamplings$get("subsampling")
r$stratify = task$target_names # stratify on target column
r$instantiate(task)
prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>       neg       pos 
#> 0.6504854 0.3495146 

prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion # FIXME why two times?

#> 
#>       neg       pos 
#> 0.6504854 0.3495146