Abstract resampling class

Abstraction for resampling strategies.

Predefined resamplings are stored in mlr_resamplings.

Usage

r = Resampling$new(id)
r$id
r$iters
r$par_set
r$par_vals
r$instantiate(task)
r$is_instantiated
r$instance
r$train_set(i)
r$test_set(i)
r$hash

Arguments

id (character(1)): identifier for this object.

i (integer(1)): Get the i-th training/test set.

Details

$new() creates a new object of class Resampling.

$id (character(1)) stores the identifier of the object.

$iters (integer(1)) calculates the resulting number of iterations, given the current par_vals.

$par_set ([paradox::ParamSet()]) describes available parameters.

$par_vals (named list) stores the currently set parameter values. You can set parameters by assigning a named list of new parameters to this slot.

$instantiate materializes fixed training and test splits for a given task.

$is_instantiated returns TRUE if the resampling has been instantiated, and FALSE otherwise.

$instance stores the instantiated realization of the resampling. This is an arbitrary object, do not work directly with it. Instead, use $train_set() and $test_set().

$train_set() returns the training set for the i-th iteration.

$test_set() returns the test set for the i-th iteration.

$hash stores a checksum (character(1)) calculated on the id, par_vals and the instantiation. If the object is not instantiated yet, NA is returned.

See also

Examples

r = mlr_resamplings$get("subsampling")

# Default parametrization
r$par_vals

#> $repeats
#> [1] 30
#> 
#> $ratio
#> [1] 0.67
#> 

# Do only 3 repeats on 10% of the data
r$par_vals = list(ratio = 0.1, repeats = 3)
r$par_vals

#> $ratio
#> [1] 0.1
#> 
#> $repeats
#> [1] 3
#> 

# Instantiate on iris task
task = mlr_tasks$get("iris")
r$instantiate(task)

#> Instantiated <ResamplingSubsampling> with 3 iterations
#> Parameters: ratio=0.1, repeats=3
#> 
#> Public: clone, has_duplicates, hash, id, instance, instantiate,
#>   is_instantiated, iters, par_set, par_vals, test_set, train_set

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>  [1]  14  21  23  33  46  48  55  72  88  90  94  98 102 125 144

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = mlr_resamplings$get("cv")$instantiate(task)
r$train_set(1)

#>   [1]   1   3   5  17  32  42  45  59  69  73  75  89 119 141 146   2  11  16
#>  [19]  22  27  47  51  77  93 101 104 112 113 116 126  12  20  41  57  61  65
#>  [37]  68  79  92  96 107 109 117 134 135   7   9  15  21  46  49  86  94  97
#>  [55]  99 114 118 130 136 142  18  24  39  55  56  71  72  80  81 125 133 139
#>  [73] 140 145 147   6   8  23  28  29  30  64  66  78  82  85 122 124 143 150
#>  [91]  10  33  34  35  40  44  63  87  88  90 102 108 115 129 144  19  31  37
#> [109]  38  43  70  76  83 105 111 120 121 132 148 149   4  14  25  48  52  54
#> [127]  60  62  67  84 100 123 131 137 138