Abstract resampling class
Source: R/Resampling.R

Abstraction for resampling strategies.

Predefined resamplings are stored in mlr_resamplings.

Usage

# Construction
r = Resampling$new(id)
#
r$id
r$iters
r$param_set
r$param_vals
r$instantiate(task)
r$is_instantiated
r$instance
r$stratify
r$train_set(i)
r$test_set(i)
r$hash

Arguments

id (character(1)):
identifier for this object.

i (integer(1)):
Get the i-th training/test set.

Details

$new() creates a new object of class Resampling.

$id (character(1)) stores the identifier of the object.

$iters (integer(1)) calculates the resulting number of iterations, given the current param_vals.

$param_set (paradox::ParamSet) describes available parameters.

$param_vals (named list) stores the currently set parameter values. You can set parameters by assigning a named list of new parameters to this slot.

$instantiate materializes fixed training and test splits for a given task.

$is_instantiated returns TRUE if the resampling has been instantiated, and FALSE otherwise.

$instance stores the instantiated realization of the resampling. This is an arbitrary object, do not work directly with it. Instead, use $train_set() and $test_set().

$stratify can be set to column names of the Task which will be used for stratification during instantiation.

$train_set() returns the training set for the i-th iteration.

$test_set() returns the test set for the i-th iteration.

$hash (character(1)) stores a checksum calculated on the id, param_vals and the instantiation. If the object is not instantiated yet, NA is returned.

See also

Examples

r = mlr_resamplings$get("subsampling")

# Default parametrization
r$param_vals

#> $repeats
#> [1] 30
#> 
#> $ratio
#> [1] 0.67
#> 

# Do only 3 repeats on 10% of the data
r$param_vals = list(ratio = 0.1, repeats = 3)
r$param_vals

#> $ratio
#> [1] 0.1
#> 
#> $repeats
#> [1] 3
#> 

# Instantiate on iris task
task = mlr_tasks$get("iris")
r$instantiate(task)

#> Instantiated <ResamplingSubsampling> with 3 iterations
#> Parameters: ratio=0.1, repeats=3
#> 
#> Public: clone, has_duplicates, hash, id, instance, instantiate,
#>   is_instantiated, iters, param_set, param_vals, stratify, test_set,
#>   train_set

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>  [1]   3   7  13  27  56  70  74  89  98 103 108 114 128 133 148

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = mlr_resamplings$get("cv")$instantiate(task)
r$train_set(1)

#>   [1]   2   7  13  26  31  34  50  53  59  77  83  94 123 143 146  20  21  70
#>  [19]  72  84  87  90  96 102 110 111 118 127 132 147   3   5   8   9  18  22
#>  [37]  47  54  58  68  89  95  98 129 142   1  24  25  33  37  49  73  76  79
#>  [55]  85 104 107 116 124 134   4  56  65  74  80  81  91 103 119 122 125 126
#>  [73] 130 138 145  11  17  23  28  43  57  61  82  93  99 100 109 114 135 144
#>  [91]  10  15  48  52  60  62  75  78  97 108 113 128 133 136 140  16  29  30
#> [109]  38  40  41  63  64  71 105 106 115 131 139 148  19  32  35  36  44  51
#> [127]  55  69  86  88 112 117 121 141 149

# Stratification
task = mlr_tasks$get("pima")
prop.table(table(task$truth())) # moderately unbalanced

#> 
#>       neg       pos 
#> 0.6510417 0.3489583 

r = mlr_resamplings$get("subsampling")
r$stratify = task$target_names # stratify on target column
r$instantiate(task)

#> Instantiated <ResamplingSubsampling> with 30 iterations
#> Parameters: repeats=30, ratio=0.67
#> 
#> Public: clone, has_duplicates, hash, id, instance, instantiate,
#>   is_instantiated, iters, param_set, param_vals, stratify, test_set,
#>   train_set

prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>      neg      pos 
#> 0.651751 0.348249 

prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>      neg      pos 
#> 0.651751 0.348249