Resampling Class

This is the abstract base class for resampling objects like ResamplingCV and ResamplingBootstrap.

The objects of this class define how a task is partitioned for resampling (e.g., in resample() or benchmark()), using a set of hyperparameters such as the number of folds in cross-validation.

Resampling objects can be instantiated on a Task, which applies the strategy on the task and manifests in a fixed partition of row_ids of the Task.

Predefined resamplings are stored in the dictionary mlr_resamplings, e.g. cv or bootstrap.

Stratification

All derived classes support stratified sampling. The stratification variables are assumed to be discrete and must be stored in the Task with column role "stratum". In case of multiple stratification variables, each combination of the values of the stratification variables forms a strata.

First, the observations are divided into subpopulations based one or multiple stratification variables (assumed to be discrete), c.f. task$strata.

Second, the sampling is performed in each of the k subpopulations separately. Each subgroup is divided into iter training sets and iter test sets by the derived Resampling. These sets are merged based on their iteration number: all training sets from all subpopulations with iteration 1 are combined, then all training sets with iteration 2, and so on. Same is done for all test sets. The merged sets can be accessed via $train_set(i) and $test_set(i), respectively.

Grouping / Blocking

All derived classes support grouping of observations. The grouping variable is assumed to be discrete and must be stored in the Task with column role "group".

Observations in the same group are treated like a "block" of observations which must be kept together. These observations either all go together into the training set or together into the test set.

The sampling is performed by the derived Resampling on the grouping variable. Next, the grouping information is replaced with the respective row ids to generate training and test sets. The sets can be accessed via $train_set(i) and $test_set(i), respectively.

See also

Public fields

id

(character(1))
Identifier of the object. Used in tables, plot and text output.

param_set

(paradox::ParamSet)
Set of hyperparameters.

instance

(any)
During instantiate(), the instance is stored in this slot in an arbitrary format. Note that if a grouping variable is present in the Task, a Resampling may operate on the group ids internally instead of the row ids (which may lead to confusion).It is advised to not work directly with the instance, but instead only use the getters $train_set() and $test_set().

task_hash

(character(1))
The hash of the Task which was passed to r$instantiate().

task_nrow

(integer(1))
The number of observations of the Task which was passed to r$instantiate().

duplicated_ids

(logical(1))
If TRUE, duplicated rows can occur within a single training set or within a single test set. E.g., this is TRUE for Bootstrap, and FALSE for cross-validation. Only used internally.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. Defaults to NA, but can be set by child classes.

Active bindings

is_instantiated

(logical(1))
Is TRUE if the resampling has been instantiated.

hash

(character(1))
Hash (unique identifier) for this object.

Methods

Public methods

Resampling$new()

Resampling$format()

Resampling$print()

Resampling$help()

Resampling$instantiate()

Resampling$train_set()

Resampling$test_set()

Resampling$clone()

Method new()

Creates a new instance of this R6 class.

Usage

Resampling$new(
  id,
  param_set = ps(),
  duplicated_ids = FALSE,
  man = NA_character_
)

Arguments

id

(character(1))
Identifier for the new instance.

param_set

(paradox::ParamSet)
Set of hyperparameters.

duplicated_ids

(logical(1))
Set to TRUE if this resampling strategy may have duplicated row ids in a single training set or test set.Note that this object is typically constructed via a derived classes, e.g. ResamplingCV or ResamplingHoldout.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method format()

Helper for print outputs.

Usage

Resampling$format()

Method print()

Printer.

Usage

Resampling$print(...)

Arguments

...

(ignored).

Method help()

Opens the corresponding help page referenced by field $man.

Usage

Resampling$help()

Method instantiate()

Materializes fixed training and test splits for a given task and stores them in r$instance in an arbitrary format.

Usage

Resampling$instantiate(task)

Arguments

task

(Task)
Task used for instantiation.

Returns

Returns the object itself, but modified by reference. You need to explicitly $clone() the object beforehand if you want to keeps the object in its previous state.

Method train_set()

Returns the row ids of the i-th training set.

Usage

Resampling$train_set(i)

Arguments

i

(integer(1))
Iteration.

Returns

(integer()) of row ids.

Method test_set()

Returns the row ids of the i-th test set.

Usage

Resampling$test_set(i)

Arguments

i

(integer(1))
Iteration.

Returns

(integer()) of row ids.

Method clone()

The objects of this class are cloneable with this method.

Usage

Resampling$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

r = rsmp("subsampling")

# Default parametrization
r$param_set$values

#> $repeats
#> [1] 30
#> 
#> $ratio
#> [1] 0.6666667
#> 

# Do only 3 repeats on 10% of the data
r$param_set$values = list(ratio = 0.1, repeats = 3)
r$param_set$values

#> $ratio
#> [1] 0.1
#> 
#> $repeats
#> [1] 3
#> 

# Instantiate on penguins task
task = tsk("penguins")
r$instantiate(task)

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>  [1]  67  44 161 258 302 263 209 109 147  97 261   2 314 295 338 237 308 189 123
#> [20] 319  64 290  68  72 169 200 331  40  53 243  88 185 139 155

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = rsmp("cv")$instantiate(task)
r$train_set(1)

#>   [1]  16  24  29  46  55  57  71  72  85  91  92 100 121 131 146 151 153 167
#>  [19] 177 179 187 192 199 205 210 213 229 241 260 275 279 283 314 322 336  14
#>  [37]  17  31  42  48  50  58  78  80  83  98 103 107 115 119 122 135 144 174
#>  [55] 178 189 191 196 207 215 225 249 258 278 285 289 291 296 312 334   6  15
#>  [73]  23  28  40  43  65  74  84  87 105 106 110 125 138 149 159 169 170 184
#>  [91] 185 186 242 244 266 281 282 297 299 309 315 318 321 328 333  12  18  20
#> [109]  25  34  37  39  60  86  88  97  99 112 133 139 154 156 162 172 188 190
#> [127] 194 238 243 251 256 262 272 274 284 290 294 325 330  13  35  38  41  45
#> [145]  51  52  59  66  77  81 126 128 142 143 147 148 195 211 223 227 236 250
#> [163] 254 255 280 286 287 298 307 323 324 335 344   3  30  32  67  68  70  73
#> [181]  89 101 104 108 127 134 137 150 160 166 176 198 201 202 206 214 240 245
#> [199] 248 259 264 273 292 310 311 331 332   4  26  47  79  96 102 111 118 120
#> [217] 123 124 129 161 163 164 180 200 208 216 226 232 246 252 263 267 268 269
#> [235] 271 276 304 308 319 338 340   5   9  11  44  75  90  94 109 113 114 117
#> [253] 130 132 141 145 158 173 183 203 204 219 220 224 230 257 265 295 300 306
#> [271] 316 320 327 337 341   7  10  19  22  27  36  49  56  61  62  64  69  76
#> [289]  95 116 152 168 181 182 193 197 218 221 222 239 247 261 270 277 288 305
#> [307] 317 329 339

# Stratification
task = tsk("pima")
prop.table(table(task$truth())) # moderately unbalanced

#> 
#>       pos       neg 
#> 0.3489583 0.6510417 

task$col_roles$stratum = task$target_names

r = rsmp("subsampling")
r$instantiate(task)
prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>       pos       neg 
#> 0.3496094 0.6503906