Resampling Class
Source: R/Resampling.R

Abstraction for resampling strategies. Predefined resamplings are stored in mlr_resamplings.

Format

R6Class object.

Usage

# Construction
r = Resampling$new(id, param_set, param_vals)
    # Members
r$duplicated_ids
r$hash
r$id
r$instance
r$is_instantiated
r$iters
r$param_set
r$stratify
r$task_hash
    # Methods
r$instantiate(task)
r$test_set(i)
r$train_set(i)

Arguments

id (character(1)): identifier for this object.

param_set (paradox::ParamSet): describes available parameters.

param_vals (list()): Named list of parameter values.

task (Task): Object of type Task.

i (integer(1)): Get the i-th training/test set.

Details

$duplicated_ids is TRUE if the resampling allows observations to be included multiple times in the training set. E.g., this is true for bootstrapping, but not for cross validation.

$hash (character(1)) stores a checksum calculated on the id, param_vals and the instantiation. If the object is not instantiated yet, NA is returned.

$id (character(1)) stores the identifier of the object.

$instance stores the instantiated realization of the resampling. This is an arbitrary object, do not work directly with it. Instead, use $train_set() and $test_set().

$instantiate materializes fixed training and test splits for a given task.

$is_instantiated returns TRUE if the resampling has been instantiated, and FALSE otherwise.

$iters (integer(1)) calculates the resulting number of iterations, given the current param_vals.

$new() creates a new object of class Resampling.

$param_set (paradox::ParamSet) describes available parameters.

$stratify can be set to column names of the Task which will be used for stratification during instantiation.

$task_hash stores the hash of the task for which the resampling has been instantiated.

$test_set() returns the test set for the i-th iteration.

$train_set() returns the training set for the i-th iteration.

See also

Examples

r = mlr_resamplings$get("subsampling")

# Default parametrization
r$param_set$values

#> $repeats
#> [1] 30
#> 
#> $ratio
#> [1] 0.67
#> 

# Do only 3 repeats on 10% of the data
r$param_set$values = list(ratio = 0.1, repeats = 3)
r$param_set$values

#> $ratio
#> [1] 0.1
#> 
#> $repeats
#> [1] 3
#> 

# Instantiate on iris task
task = mlr_tasks$get("iris")
r$instantiate(task)

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>  [1]   8  32  38  64  66  74  84  91  96 100 111 124 127 138 144

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = mlr_resamplings$get("cv")$instantiate(task)
r$train_set(1)

#>   [1]  10  41  53  54  82  86  91 103 115 122 128 131 134 135 150   5   9  13
#>  [19]  17  26  42  68  81  87  89 102 109 132 144 145   4   6  16  19  21  23
#>  [37]  30  56  77  99 104 105 130 141 143   2  22  44  46  57  65  72  78  80
#>  [55] 111 112 133 137 139 148  18  24  25  28  39  48  58  96  98 100 106 124
#>  [73] 142 147 149   1   7  12  32  37  47  50  62  70  76  93  94 108 117 146
#>  [91]  33  38  40  45  74  84  85  88  92 118 120 121 125 126 129  14  15  29
#> [109]  35  36  43  55  66  73 114 116 119 127 136 138  11  20  31  49  51  59
#> [127]  61  64  67  75  83  90 101 110 113

# Stratification
task = mlr_tasks$get("pima")
prop.table(table(task$truth())) # moderately unbalanced

#> 
#>       neg       pos 
#> 0.6510417 0.3489583 

r = mlr_resamplings$get("subsampling")
r$stratify = task$target_names # stratify on target column
r$instantiate(task)
prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>       neg       pos 
#> 0.6504854 0.3495146 

prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion # FIXME why two times?

#> 
#>       neg       pos 
#> 0.6504854 0.3495146