Resampling Class

This is the abstract base class for resampling objects like ResamplingCV and ResamplingBootstrap.

The objects of this class define how a task is partitioned for resampling (e.g., in resample() or benchmark()), using a set of hyperparameters such as the number of folds in cross-validation.

Resampling objects can be instantiated on a Task, which applies the strategy on the task and manifests in a fixed partition of row_ids of the Task.

Predefined resamplings are stored in the mlr3misc::Dictionary mlr_resamplings, e.g. cv or bootstrap.

Format

R6::R6Class object.

Construction

Note: This object is typically constructed via a derived classes, e.g. ResamplingCV or ResamplingHoldout.

r = Resampling$new(id, param_set, duplicated_ids = FALSE, man = NA_character_)

id :: character(1)
Identifier for the resampling strategy.

param_set :: paradox::ParamSet
Set of hyperparameters.

duplicated_ids :: logical(1)
Set to TRUE if this resampling strategy may have duplicated row ids in a single training set or test set.

man :: character(1)
String in the format [pkg]::[topic] pointing to a manual page for this object.

Fields

All variables passed to the constructor, and additionally:

iters :: integer(1)
Return the number of resampling iterations, depending on the values stored in the param_set.

instance :: any
During instantiate(), the instance is stored in this slot. The instance can be in any arbitrary format.

is_instantiated :: logical(1)
Is TRUE, if the resampling has been instantiated.

task_hash :: character(1)
The hash of the Task which was passed to r$instantiate().

task_nrow :: integer(1)
The number of observations of the Task which was passed to r$instantiate().

hash :: character(1)
Hash (unique identifier) for this object.

E.g., this is TRUE for Bootstrap, and FALSE for cross validation. Only used internally.

Methods

instantiate(task)
Task -> self
Materializes fixed training and test splits for a given task and stores them in r$instance.

train_set(i)
integer(1) -> (integer() | character())
Returns the row ids of the i-th training set.

test_set(i)
integer(1) -> (integer() | character())
Returns the row ids of the i-th test set.

help()
() -> NULL
Opens the corresponding help page referenced by $man.

Stratification

All derived classes support stratified sampling. The stratification variables are assumed to be discrete and must be stored in the Task with column role "stratum". In case of multiple stratification variables, each combination of the values of the stratification variables forms a strata.

First, the observations are divided into subpopulations based one or multiple stratification variables (assumed to be discrete), c.f. task$strata.

Second, the sampling is performed in each of the k subpopulations separately. Each subgroup is divided into iter training sets and iter test sets by the derived Resampling. These sets are merged based on their iteration number: all training sets from all subpopulations with iteration 1 are combined, then all training sets with iteration 2, and so on. Same is done for all test sets. The merged sets can be accessed via $train_set(i) and $test_set(i), respectively.

Grouping / Blocking

All derived classes support grouping of observations. The grouping variable is assumed to be discrete and must be stored in the Task with column role "group".

Observations in the same group are treated like a "block" of observations which must be kept together. These observations either all go together into the training set or together into the test set.

The sampling is performed by the derived Resampling on the grouping variable. Next, the grouping information is replaced with the respective row ids to generate training and test sets. The sets can be accessed via $train_set(i) and $test_set(i), respectively.

See also

Examples

r = rsmp("subsampling")

# Default parametrization
r$param_set$values

#> $repeats
#> [1] 30
#> 
#> $ratio
#> [1] 0.6666667
#> 

# Do only 3 repeats on 10% of the data
r$param_set$values = list(ratio = 0.1, repeats = 3)
r$param_set$values

#> $ratio
#> [1] 0.1
#> 
#> $repeats
#> [1] 3
#> 

# Instantiate on iris task
task = tsk("iris")
r$instantiate(task)

# Extract train/test sets
train_set = r$train_set(1)
print(train_set)

#>  [1]  17  65  39  21 145 115  58  81  50 124  90  95  85 114  69

intersect(train_set, r$test_set(1))

#> integer(0)

# Another example: 10-fold CV
r = rsmp("cv")$instantiate(task)
r$train_set(1)

#>   [1]   5  15  29  55  57  58  59  64  71  75  99 102 115 126 142  14  18  27
#>  [19]  47  51  70  72  81  96  97 107 112 116 123 128   7   8  10  21  23  35
#>  [37]  36  62  65  73  78  95 108 124 137   4  12  13  22  34  76  89  90  93
#>  [55] 104 131 134 135 145 149   2   6  19  20  30  31  60  84  92 117 118 132
#>  [73] 133 141 147   1   9  32  33  44  50  68  79  80  82 109 127 129 139 144
#>  [91]  16  24  37  43  46  54  67  69  87 100 119 122 130 136 148  17  25  38
#> [109]  39  41  48  61  66  85  86  98 105 114 143 150  11  26  28  49  52  74
#> [127]  77  83  88 101 103 110 111 120 146

# Stratification
task = tsk("pima")
prop.table(table(task$truth())) # moderately unbalanced

#> 
#>       pos       neg 
#> 0.3489583 0.6510417 

task$col_roles$stratum = task$target_names

r = rsmp("subsampling")
r$instantiate(task)
prop.table(table(task$truth(r$train_set(1)))) # roughly same proportion

#> 
#>       pos       neg 
#> 0.3496094 0.6503906