Supervised or Unsupervised Tasks
Source: R/Task.R

This is the abstract base class for task objects like TaskClassif and TaskRegr.

Format

R6Class object

Usage

# Construction
t = Task$new(id, backend, task_type)
    # Members
t$backend
t$col_info
t$col_roles
t$features_names
t$feature_types
t$formula
t$groups
t$hash
t$id
t$measures
t$ncol
t$nrow
t$properties
t$row_ids
t$row_roles
t$target_names
t$task_type
    # Methods: Accessors
t$data(rows = NULL, cols = NULL)
t$head(n = 6)
t$levels(col)
t$set_col_role(cols, new_roles, exclusive = TRUE)
t$set_row_role(rows, new_roles, exclusive = TRUE)
    # Methods: Mutators
t$cbind(data)
t$filter(rows)
t$overwrite(data)
t$rbind(data)
t$select(cols)

Arguments

id (character(1)): Name of the task.

backend (DataBackend): DataBackend which stores the data.

task_type (character(1)): Task type. Set via class which inherits from Task.

rows (integer() | character()): Vector of row ids specifying rows from the DataBackend using its primary key. Can be character() or integer, depending on the DataBackend.

cols (character()): Character vector to specify columns from the DataBackend.

n (integer(1)): Number of rows to retrieve from the DataBackend.

col (character(1)): Character vector to specify a single column from the DataBackend.

new_roles (character(1)): New roles to assign for specified rows/columns.

exclusive (logical(1)): If TRUE, the cols/rows will be removed from all roles except new_roles.

data (data.frame()): New data to rbind/cbind to the task.

Details

$backend (DataBackend) stores the DataBackend of the task.

$cbind extends the task with additional columns. The row ids must be provided as column in data (with column name matching the primary key name of the DataBackend).

$col_info (data.table) with columns id, and type and levels. Stores column names of DataBackend in column id. Column type stores the storage type of the variables, e.g. integer, numeric or character. Column levels stores the levels for factor and character variables.

$col_roles (list). Each column (feature) can have a specific mutually exclusive role in the learning task:

"feature": Regular feature.

"target": Column with target labels.

"order": Returned data is ordered by these column(s).

"groups": During resampling, observations with the same value of the variable listed in "groups" are marked as "belonging together". They will be assigned jointly to be either in the training set or the test set. Returns a (data.table::data.table) with two columns: first column are rows ids, second column are the group labels.

"weights": Observation weights. (data.table::data.table) with two columns: first column are the row ids, second column are the observation weights. To alter the role, use $set_col_role()

$data is used to retrieve data from the backend as data.table. Rows are subsetted to only contain observations with role == "use". Columns are filtered to only contain features with role %in% c("target", "feature"). If invalid rows or cols are specified, an exception is raised.

$feature_names (character()) returns all column names with role == "feature".

$feature_types (data.table::data.table)returns a table with columns id and type where id are the column names of "active" features of the task and type is the storage type.

$filter reduces the task, subsetting it to only the rows specified.

$formula constructs a stats::formula, e.g. [target] ~ [feature_1] + [feature_2] + ... + [feature_k].

$groups returns a (data.table::data.table) with two columns: the row ids and the grouping / blocking information.

$hash (character(1)) stores a checksum calculated on the id, row_roles and col_roles.

$head (data.table::data.table) can be used to peek into the first n observations with role == "use".

$id (character(1)) stores the name of the task.

$measures (list of Measure) stores the default measures for this task.

$ncol (integer(1)) provides the total number of cols with role %in% c("target", "feature").

$nrow (integer(1)) provides the total number of rows with role == "use".

FIXME properties missing

$row_ids (data.table()) returns the active row ids used in the backend, i.e. subsetted to observations with role == "use". The column names of the returned data.table equals the primary key column in the DataBackend.

$row_roles (list). Stores the row ids of DataBackend in vectors of row roles:

"use": Use in training.

"validation": Do not use in training, this are (possibly unlabeled) observations which are held back unless explicitly addressed. To alter the role, use set_row_role().

$target_names (character()) returns all column names with role == "target".

FIXME task_type missing

$new() initializes a new object of class Task.

$levels() (character()) queries the distinct levels of the column col. Only works for character and factor columns. This function ignores the row roles, so you get all levels found in the DataBackend.

$overwrite() overwrite the data in the DataBackend with data provided as data.table::data.table(). The row ids must be provided as column in data (with column name matching the primary key name of the DataBackend).

$rbind() extends the task with additional rows.

$select() reduces the task, subsetting it to only the columns specified.

$set_col_role() sets the role for specified columns, referenced by name. If exclusive is TRUE, the referenced columns will be removed from all other roles.

$set_row_role() sets the role for specified rows, referenced by row id. If exclusive is TRUE, the referenced rows will be removed from all other roles.

Task mutators

The methods filter(), select(), rbind(), cbind(), and overwrite() change the task in-place, but without modifying the DataBackend. filter() and select() just reduce the set of active rows or columns, providing a different view on the data. rbind(), cbind(), and overwrite() first create a new DataBackendDataTable from the provided data, and then merge both backends into an abstract DataBackend which combines the results on-demand.

References

HTML help page

See also

Examples

b = as_data_backend(iris)
task = Task$new("iris", task_type = "classif", backend = b)

task$nrow

#> [1] 150

task$ncol

#> [1] 5

task$head()

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width Species
#> 1:          1.4         0.2          5.1         3.5  setosa
#> 2:          1.4         0.2          4.9         3.0  setosa
#> 3:          1.3         0.2          4.7         3.2  setosa
#> 4:          1.5         0.2          4.6         3.1  setosa
#> 5:          1.4         0.2          5.0         3.6  setosa
#> 6:          1.7         0.4          5.4         3.9  setosa

task$feature_names

#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width"  "Species"     

task$formula

#> ~Petal.Length + Petal.Width + Sepal.Length + Sepal.Width + Species
#> NULL

# Remove "Petal.Length"
task$set_col_role("Petal.Length", character(0L))

# Remove "Petal.Width", alternative way
task$select(setdiff(task$feature_names, "Petal.Width"))

task$feature_names

#> [1] "Sepal.Length" "Sepal.Width"  "Species"     

# Add new column "foo"
task$cbind(cbind(data.frame(foo = 1:150), task$row_ids))