Abstract learning task
Source: R/Task.R

This is the abstract base class for task objects like TaskClassif and TaskRegr.

Usage

# Construction
t = Task$new(id, backend)
#
t$id
t$backend
t$row_roles
t$col_info
t$set_row_role(rows, new_roles, exclusive = TRUE)
t$set_col_role(cols, new_roles, exclusive = TRUE)
t$measures
t$data(rows = NULL, cols = NULL)
t$head(n = 6)
t$levels(col)
t$row_ids(subset = NULL)
t$features_names
t$target_names
t$nrow
t$ncol
t$feature_types
t$formula
t$hash
#
t$filter(rows)
t$select(cols)
t$rbind(data)
t$cbind(data)
t$overwrite(cols)

Arguments

id (string): Name of the task.

backend (DataBackend): DataBackend which stores the data.

data (base::data.frame): New data to rbind/cbind to the task.

rows (vector): Vector of row ids specifying rows from the DataBackend using its primary key. Can be character() or integer, depending on the DataBackend.

cols (character()): Character vector to specify columns from the DataBackend.

col (character(1)): Character vector to specify a single column from the DataBackend.

n (integer(1)): Number of rows to retrieve from the DataBackend.

new_roles (character(1)): New roles to assign for specified rows/columns.

exclusive (logical(1)): If TRUE, the cols/rows will be removed from all roles except new_roles.

subset (vector): Subset of row ids to subset rows from the DataBackend using its primary key.

Details

$new() initializes a new object of class Task.

$id (character(1)) stores the name of the task.

$backend() (DataBackend) stores the DataBackend of the task.

$row_roles (list). Stores the row ids of DataBackend in vectors of row roles:

"use": Use in training.

"validation": Do not use in training, this are (possibly unlabeled) observations which are held back unless explicitly addressed. To alter the role, use set_row_role().

$col_info (data.table) with columns id, role and type. Stores column names of DataBackend in column id. Each column (feature) can have a specific mutually exclusive role in the learning task:

"feature": Regular feature.

"target": Column with target labels.

"ignore": Do not these features at all.

"primary_key": Name of the primary id column used in DataBackend. Column type stores the storage type of the variable, e.g. integer, numeric or character. To alter the role, use set_col_role().

$set_row_role() overwrites the role for specified rows, referenced by row id.

$set_col_role() overwrites the role for specified columns.

$measures is a list of Measure (performance measures) to use in this task.

$data() is used to retrieve data from the backend as data.table. Rows are subsetted to only contain observations with role == "use". Columns are filtered to only contain features with role %in% c("target", "feature"). If invalid rows or cols are specified, an exception is raised.

$head() can be used to peek into the first n observations with role == "use".

$levels() queries the distinct levels of the column col. Only works for character and factor columns. This function ignores the row roles, so you get all levels found in the DataBackend.

$row_ids() returns a (subset of) row ids used in the task, i.e. subsetted to observations with role == "use".

$feature_names returns a character vector of all feature names with role == "feature".

$target_names returns a character vector of all feature names with role == "target".

$nrow provides the total number of rows with role == "use".

$ncol provides the total number of cols with role %in% c("target", "feature").

$feature_types gives a data.table with columns id and type where id are the column names of "active" features of the task and type is the storage type.

$formula constructs a stats::formula, e.g. [target] ~ [feature_1] + [feature_2] + ... + [feature_k].

$filter() reduces the task, subsetting it to only the rows specified.

$select() reduces the task, subsetting it to only the columns specified.

$rbind() extends the task with additional rows.

$cbind() extends the task with additional columns.

$overwrite() overwrite the data in the DataBackend with data provided as data.table(). Values to overwrite are matched via column names and primary key.

$hash stores a checksum (character(1)) calculated on the id, row_roles and col_info.

See also

Examples

b = DataBackendDataTable$new(iris)
task = Task$new("iris", b)
task$nrow

#> [1] 150

task$ncol

#> [1] 5

task$head()

#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1:          5.1         3.5          1.4         0.2  setosa
#> 2:          4.9         3.0          1.4         0.2  setosa
#> 3:          4.7         3.2          1.3         0.2  setosa
#> 4:          4.6         3.1          1.5         0.2  setosa
#> 5:          5.0         3.6          1.4         0.2  setosa
#> 6:          5.4         3.9          1.7         0.4  setosa

task$formula

#> ~Sepal.Length + Sepal.Width + Petal.Length + Petal.Width + Species
#> NULL

# Remove "Petal.Length"
task$set_col_role("Petal.Length", character(0L))
task$formula

#> ~Sepal.Length + Sepal.Width + Petal.Width + Species
#> NULL