Task Class
Source: R/Task.R

This is the abstract base class for task objects like TaskClassif and TaskRegr. Predefined tasks are stored in mlr_tasks.

Format

R6::R6Class object.

Construction

t = Task$new(id, task_type, backend)

id :: character(1)
Name of the task.

task_type :: character(1)
Set in the classes which inherit from this class. Must be an element of mlr_reflections$task_types.

backend :: DataBackend

Fields

backend :: DataBackend.

col_info :: data.table::data.table()
Table with with 3 columns: Column names of DataBackend are stored in columnid. Column type holds the storage type of the variables, e.g. integer, numeric or character. Column levels keeps a list of possible levels for factor and character variables.

col_roles :: named list()
Each column (feature) can have an arbitrary number of roles in the learning task:

"target": Labels to predict.

"feature": Regular feature.

"order": Data returned by data() is ordered by this column (or these columns).

"groups": During resampling, observations with the same value of the variable with role "groups" are marked as "belonging together". They will be exclusively assigned to be either in the training set or the test set for each resampling iteration. Only a single column may be marked as grouping column.

"weights": Observation weights. Only a single column may be marked as weights. col_roles keeps track of the roles with a named list of vectors of feature names. To alter the roles, use t$set_col_role().

row_roles :: named list()
Each row (observation) can have an arbitrary number of roles in the learning task:

"use": Use in train / predict / resampling.

"validation": Hold the observations back unless explicitly requested. row_roles keeps track of the roles with a named list of vectors of feature names. To alter the role, use set_row_role().

feature_names :: character()
Returns all column names with role == "feature".

feature_types :: data.table::data.table()
Returns a table with columns id and type where id are the column names of "active" features of the task and type is the storage type.

formula :: formula()
Constructs a stats::formula, e.g. [target] ~ [feature_1] + [feature_2] + ... + [feature_k], using the active features of the task.

hash :: character(1)
Hash (unique identifier) for this object.

id :: character(1)
Stores the identifier of the Task.

measures :: list() of Measure
Stores the measures to use for this task.

ncol :: integer(1)
Returns the total number of cols with role "target" or "feature".

nrow :: integer(1)
Return the total number of rows with role "use".

row_ids :: (integer() | character())
Returns the row ids of the DataBackend for observations with with role "use".

target_names :: character()
Returns all column names with role "target".

task_type :: character(1)
Stores the type of the Task.

groups :: data.table::data.table()
If the task has a designated column role "groups", table with two columns: "row_id" (integer() | character()) and the grouping variable group (vector()). Returns NULL if there are is no grouping column.

weights :: data.table::data.table()
If the task has a designated column role "weights", table with two columns: "row_id" (integer() | character()) and the observation weights weight (numeric()). Returns NULL if there are is no weight column.

Methods

data(rows = NULL, cols = NULL, format = NULL)
(integer() | character(), character(), character(1)) -> any
Returns a slice of the data from the DataBackend in the format specified by format (depending on the DataBackend, but usually a data.table::data.table()). Rows are subsetted to only contain observations with role "use". Columns are filtered to only contain features with roles "target" and "feature". If invalid rows or cols are specified, an exception is raised.

cbind(data)
data.frame() -> self
Extends the DataBackend with additional columns. The row ids must be provided as column in data (with column name matching the primary key name of the DataBackend). If this column is missing, it is assumed that the rows are exactly in the order of t$row_ids.

rbind(data)
data.frame() -> self
Extends the DataBackend with additional rows. The new row ids must be provided as column in data. If this column is missing, new row ids are constructed automatically.

filter(rows)
(integer() | character()) -> self
Subsets the task, reducing it to only keep the rows specified.

select(cols)
character() -> self
Subsets the task, reducing it to only keep the columns specified.

levels(col)
character() -> named list()
Returns the distinct levels of the column col. Only applicable for features with type "character", "factor" or "ordered". This function ignores the row roles, it returns all levels available in the DataBackend.

head(n = 6)
integer() -> data.table::data.table()
Get the first n observations with role "use".

replace_features(data)
data.frame() -> self
Replaces some features of the task by constructing a completely new DataBackendDataTable. This operation is similar to calling select() and cbind(), but explicitly copies the data.

set_col_role(cols, new_roles, exclusive = TRUE)
(character(), character(), logical(1)) -> self
Adds the roles new_roles to columns referred to by cols. If exclusive is TRUE, the referenced columns will be removed from all other roles.

set_row_role(rows, new_roles, exclusive = TRUE)
(character(), character(), logical(1)) -> self
Adds the roles new_roles to rows referred to by rows. If exclusive is TRUE, the referenced rows will be removed from all other roles.

S3 methods

as.data.frame(task)
Task -> data.frame()
Returns the data set as data.frame().

as.data.table(task)
Task -> data.table::data.table()
Returns the data set as data.table().

Task mutators

The following methods change the task in-place:

set_row_roles() and set_col_roles() alter the row or column information in row_roles or col_roles, respectively.

filter() and select() subset the set of active rows or columns in row_roles or col_roles, respectively. This provides a different "view" on the data.

rbind() and cbind() change the task in-place by binding rows or columns to the data, but without modifying the original DataBackend. Instead, the methods first create a new DataBackendDataTable from the provided new data, and then merge both backends into an abstract DataBackend which combines the results on-demand.

replace_features() is a convenience wrapper around select() and cbind(). Again, the original DataBackend remains unchanged.

See also

Examples

b = as_data_backend(iris)
task = Task$new("iris", task_type = "classif", backend = b)

task$nrow

#> [1] 150

task$ncol

#> [1] 5

task$head()

#>    Petal.Length Petal.Width Sepal.Length Sepal.Width Species
#> 1:          1.4         0.2          5.1         3.5  setosa
#> 2:          1.4         0.2          4.9         3.0  setosa
#> 3:          1.3         0.2          4.7         3.2  setosa
#> 4:          1.5         0.2          4.6         3.1  setosa
#> 5:          1.4         0.2          5.0         3.6  setosa
#> 6:          1.7         0.4          5.4         3.9  setosa

task$feature_names

#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width"  "Species"     
#> attr(,".unique")
#> [1] TRUE

task$formula

#> ~Petal.Length + Petal.Width + Sepal.Length + Sepal.Width + Species
#> NULL

# Remove "Petal.Length"
task$set_col_role("Petal.Length", character(0L))

# Remove "Petal.Width", alternative way
task$select(setdiff(task$feature_names, "Petal.Width"))

task$feature_names

#> [1] "Sepal.Length" "Sepal.Width"  "Species"     
#> attr(,".unique")
#> [1] TRUE

# Add new column "foo"
task$cbind(data.frame(foo = 1:150))