TaskGenerator Class

Creates a Task of arbitrary size. Predefined task generators are stored in the dictionary mlr_task_generators, e.g. xor.

See also

Public fields

id

(character(1))
Identifier of the object. Used in tables, plot and text output.

label

(character(1))
Label for this object. Can be used in tables, plot and text output instead of the ID.

task_type

(character(1))
Task type, e.g. "classif" or "regr".

For a complete list of possible task types (depending on the loaded packages), see mlr_reflections$task_types$type.

param_set

(paradox::ParamSet)
Set of hyperparameters.

packages

(character(1))
Set of required packages. These packages are loaded, but not attached.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. Defaults to NA, but can be set by child classes.

Methods

Public methods

TaskGenerator$new()

TaskGenerator$format()

TaskGenerator$print()

TaskGenerator$generate()

TaskGenerator$clone()

Method new()

Creates a new instance of this R6 class.

Usage

TaskGenerator$new(
  id,
  task_type,
  packages = character(),
  param_set = ps(),
  label = NA_character_,
  man = NA_character_
)

Arguments

id

(character(1))
Identifier for the new instance.

task_type

(character(1))
Type of task, e.g. "regr" or "classif". Must be an element of mlr_reflections$task_types$type.

packages

(character())
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not installed, but loaded (not attached) later on-demand via requireNamespace().

param_set

(paradox::ParamSet)
Set of hyperparameters.

label

(character(1))
Label for the new instance.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method format()

Helper for print outputs.

Usage

TaskGenerator$format(...)

Arguments

...

(ignored).

Method print()

Printer.

Usage

TaskGenerator$print(...)

Arguments

...

(ignored).

Method generate()

Creates a task of type task_type with n observations, possibly using additional settings stored in param_set.

Usage

TaskGenerator$generate(n)

Arguments

n

(integer(1))
Number of rows to generate.

Returns

Task.

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskGenerator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.