TaskGenerator Class

Creates a Task of arbitrary size. Predefined task generators are stored in the mlr3misc::Dictionary mlr_task_generators, e.g. xor.

Format

R6::R6Class object.

Construction

g = TaskGenerator$new(id, task_type, packages = character(), param_set = ParamSet$new(), man = NA_character_)

id :: character(1)
Identifier for the learner.

task_type :: character(1)
Type of the task the learner can operator on. E.g., "classif" or "regr".

packages :: character()
Set of required packages. Note that these packages will be loaded via requireNamespace(), and are not attached.

param_set :: paradox::ParamSet
Set of hyperparameters.

man :: character(1)
String in the format [pkg]::[topic] pointing to a manual page for this object.

Fields

All variables passed to the constructor, and additionally:

task_type :: character(1)
Stores the type of class this learner can operate on, e.g. "classif" or "regr". A complete list of task types is stored in mlr_reflections$task_types$type.

Methods

generate(n)
integer(1) -> Task
Creates a task of type task_type with n observations, possibly using additional settings stored in param_set.

help()
() -> NULL
Opens the corresponding help page referenced by $man.

See also