Supervised Task

This is the abstract base class for task objects like TaskClassif and TaskRegr. It extends Task with methods to handle a target columns. Supervised tasks for probabilistic regression (including survival analysis) can be found in mlr3proba.

See also

Super class

mlr3::Task -> TaskSupervised

Methods

Public methods

TaskSupervised$new()

TaskSupervised$truth()

TaskSupervised$clone()

Inherited methods

mlr3::Task$add_strata()

mlr3::Task$cbind()

mlr3::Task$data()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$print()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_roles()

mlr3::Task$set_levels()

mlr3::Task$set_row_roles()

Method new()

Creates a new instance of this R6 class.

Usage

TaskSupervised$new(
  id,
  task_type,
  backend,
  target,
  label = NA_character_,
  extra_args = list()
)

Arguments

id

(character(1))
Identifier for the new instance.

task_type

(character(1))
Type of task, e.g. "regr" or "classif". Must be an element of mlr_reflections$task_types$type.

backend

(DataBackend)
Either a DataBackend, or any object which is convertible to a DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a DataBackendDataTable.

target

(character(1))
Name of the target column.

label

(character(1))
Label for the new instance.

extra_args

(named list())
Named list of constructor arguments, required for converting task types via convert_task().

Method truth()

True response for specified row_ids. Format depends on the task type. Defaults to all rows with role "use".

Usage

TaskSupervised$truth(rows = NULL)

Arguments

rows

integer()
Row indices.

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskSupervised$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

TaskSupervised$new("penguins", task_type = "classif", backend = palmerpenguins::penguins,
  target = "species")
#> <TaskSupervised:penguins> (344 x 8)
#> * Target: species
#> * Properties: -
#> * Features (7):
#>   - int (3): body_mass_g, flipper_length_mm, year
#>   - dbl (2): bill_depth_mm, bill_length_mm
#>   - fct (2): island, sex