Supervised Task

This is the abstract base class for task objects like TaskClassif and TaskRegr. It extends Task with methods to handle a target columns.

Format

R6::R6Class object inheriting from Task.

Construction

t = TaskSupervised$new(id, task_type, backend, target)

id :: character(1)
Identifier for the task.

backend :: DataBackend
Either a DataBackend, or any object which is convertible to a DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a DataBackendDataTable.

task_type :: character(1)
Set in the classes which inherit from this class. Must be an element of mlr_reflections$task_types$type.

target :: character(1)
Name of the target column.

Fields

See Task.

Methods

All methods from Task, and additionally:

truth(rows = NULL) :: any
True response for specified row_ids. Format depends on the task type. Defaults to all rows with role "use".

See also

Examples

task = TaskSupervised$new("iris", task_type = "classif", backend = iris, target = "Species")