Unsupervised Task

This is the abstract base class for unsupervised tasks such as cluster tasks in mlr3cluster and mlr3spatial.

See also

Super class

mlr3::Task -> TaskUnsupervised

Methods

Public methods

TaskUnsupervised$new()

TaskUnsupervised$clone()

Inherited methods

mlr3::Task$add_strata()

mlr3::Task$cbind()

mlr3::Task$data()

mlr3::Task$divide()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$print()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_roles()

mlr3::Task$set_levels()

mlr3::Task$set_row_roles()

Method new()

Creates a new instance of this R6 class.

Usage

TaskUnsupervised$new(
  id,
  task_type = "unsupervised",
  backend,
  label = NA_character_,
  extra_args = list()
)

Arguments

id

(character(1))
Identifier for the new instance.

task_type

(character(1))
Type of task, e.g. "regr" or "classif". Must be an element of mlr_reflections$task_types$type.

backend

(DataBackend)
Either a DataBackend, or any object which is convertible to a DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a DataBackendDataTable.

label

(character(1))
Label for the new instance.

extra_args

(named list())
Named list of constructor arguments, required for converting task types via convert_task().

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskUnsupervised$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

TaskUnsupervised$new("penguins", task_type = "regr", backend = palmerpenguins::penguins)
#> 
#> ── <TaskUnsupervised> (344x8) ──────────────────────────────────────────────────
#> • Target:
#> • Properties: -
#> • Features (8):
#>   • int (3): body_mass_g, flipper_length_mm, year
#>   • fct (3): island, sex, species
#>   • dbl (2): bill_depth_mm, bill_length_mm