Unsupervised Task

This is the abstract base class for unsupervised tasks such as cluster tasks (not yet implemented).

See also

Super class

mlr3::Task -> TaskUnsupervised

Methods

Public methods

TaskUnsupervised$clone()

Inherited methods

mlr3::Task$cbind()

mlr3::Task$data()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$initialize()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$print()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_role()

mlr3::Task$set_row_role()

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskUnsupervised$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

TaskUnsupervised$new("iris", task_type = "regr", backend = iris)

#> <TaskUnsupervised:iris> (150 x 5)
#> * Target: -
#> * Properties: -
#> * Features (5):
#>   - dbl (4): Petal.Length, Petal.Width, Sepal.Length, Sepal.Width
#>   - fct (1): Species