Manually construct an object of type ResultData

This function allows to manually construct a ResampleResult or BenchmarkResult by combining the individual components to an object of class ResultData, mlr3's internal container object. A ResampleResult or BenchmarkResult can then be initialized with the returned object. Note that ResampleResults can be converted to a BenchmarkResult with as_benchmark_result() and multiple BenchmarkResults can be combined to a larger BenchmarkResult.

as_result_data(
  task,
  learners,
  resampling,
  iterations,
  predictions,
  learner_states = NULL
)

Arguments

task
(Task).

learners
(list of trained Learners).

resampling
(Resampling).

iterations
(integer()).

predictions
(list of Predictions).

learner_states
(list())
Learner states. If not provided, the states of learners are automatically extracted.

Value

ResultData object which can be passed to the constructor of ResampleResult.

Examples

task = tsk("iris")
learner = lrn("classif.rpart")
resampling = rsmp("cv", folds = 2)$instantiate(task)
iterations = seq_len(resampling$iters)

# manually train two learners.
# store learners and predictions
learners = list()
predictions = list()
for (i in iterations) {
  l = learner$clone(deep = TRUE)
  learners[[i]] = l$train(task, row_ids = resampling$train_set(i))
  predictions[[i]] = l$predict(task, row_ids = resampling$test_set(i))
}

rdata = as_result_data(task, learners, resampling, iterations, predictions)
ResampleResult$new(rdata)

#> <ResampleResult> of 2 iterations
#> * Task: iris
#> * Learner: classif.rpart
#> * Warnings: 0 in 0 iterations
#> * Errors: 0 in 0 iterations