Generate a Benchmark Grid Design

Takes a lists of Task, a list of Learner and a list of Resampling to generate a design in an expand.grid() fashion (a.k.a. cross join or Cartesian product).

Resampling strategies are not allowed to be instantiated when passing the argument, and instead will be instantiated per task internally.

benchmark_grid(tasks, learners, resamplings)

Arguments

tasks
:: list of Task.

learners
:: list of Learner.

resamplings
:: list of Resampling.

Value

(data.table::data.table()) with the cross product of the input vectors.

Examples

tasks = list(tsk("iris"), tsk("sonar"))
learners = list(lrn("classif.featureless"), lrn("classif.rpart"))
resamplings = list(rsmp("cv"), rsmp("subsampling"))
benchmark_grid(tasks, learners, resamplings)

#>             task                     learner              resampling
#> 1: <TaskClassif> <LearnerClassifFeatureless>          <ResamplingCV>
#> 2: <TaskClassif> <LearnerClassifFeatureless> <ResamplingSubsampling>
#> 3: <TaskClassif>       <LearnerClassifRpart>          <ResamplingCV>
#> 4: <TaskClassif>       <LearnerClassifRpart> <ResamplingSubsampling>
#> 5: <TaskClassif> <LearnerClassifFeatureless>          <ResamplingCV>
#> 6: <TaskClassif> <LearnerClassifFeatureless> <ResamplingSubsampling>
#> 7: <TaskClassif>       <LearnerClassifRpart>          <ResamplingCV>
#> 8: <TaskClassif>       <LearnerClassifRpart> <ResamplingSubsampling>