Generate a Benchmark Grid Design

Takes a lists of Task, a list of Learner and a list of Resampling to generate a design in an expand.grid() fashion (a.k.a. cross join or Cartesian product).

Resampling strategies are not allowed to be instantiated when passing the argument, and instead will be instantiated per task internally. The only exception to this rule applies if all tasks have exactly the same number of rows, and the resamplings are all instantiated for such tasks.

benchmark_grid(tasks, learners, resamplings)

Arguments

tasks

(list of Task).

learners

(list of Learner).

resamplings

(list of Resampling).

Value

(data.table::data.table()) with the cross product of the input vectors.

See also

Examples

tasks = list(tsk("penguins"), tsk("sonar"))
learners = list(lrn("classif.featureless"), lrn("classif.rpart"))
resamplings = list(rsmp("cv"), rsmp("subsampling"))

grid = benchmark_grid(tasks, learners, resamplings)
print(grid)
#>                 task                         learner
#> 1: <TaskClassif[49]> <LearnerClassifFeatureless[37]>
#> 2: <TaskClassif[49]> <LearnerClassifFeatureless[37]>
#> 3: <TaskClassif[49]>       <LearnerClassifRpart[37]>
#> 4: <TaskClassif[49]>       <LearnerClassifRpart[37]>
#> 5: <TaskClassif[49]> <LearnerClassifFeatureless[37]>
#> 6: <TaskClassif[49]> <LearnerClassifFeatureless[37]>
#> 7: <TaskClassif[49]>       <LearnerClassifRpart[37]>
#> 8: <TaskClassif[49]>       <LearnerClassifRpart[37]>
#>                     resampling
#> 1:          <ResamplingCV[19]>
#> 2: <ResamplingSubsampling[19]>
#> 3:          <ResamplingCV[19]>
#> 4: <ResamplingSubsampling[19]>
#> 5:          <ResamplingCV[19]>
#> 6: <ResamplingSubsampling[19]>
#> 7:          <ResamplingCV[19]>
#> 8: <ResamplingSubsampling[19]>
if (FALSE) {
benchmark(grid)
}

# manual construction of the grid with data.table::CJ()
grid = data.table::CJ(task = tasks, learner = learners,
  resampling = resamplings, sorted = FALSE)

# manual instantiation (not suited for a fair comparison of learners!)
Map(function(task, resampling) {
  resampling$instantiate(task)
}, task = grid$task, resampling = grid$resampling)
#> [[1]]
#> <ResamplingCV> with 10 iterations
#> * Instantiated: TRUE
#> * Parameters: folds=10
#> 
#> [[2]]
#> <ResamplingSubsampling> with 30 iterations
#> * Instantiated: TRUE
#> * Parameters: repeats=30, ratio=0.6667
#> 
#> [[3]]
#> <ResamplingCV> with 10 iterations
#> * Instantiated: TRUE
#> * Parameters: folds=10
#> 
#> [[4]]
#> <ResamplingSubsampling> with 30 iterations
#> * Instantiated: TRUE
#> * Parameters: repeats=30, ratio=0.6667
#> 
#> [[5]]
#> <ResamplingCV> with 10 iterations
#> * Instantiated: TRUE
#> * Parameters: folds=10
#> 
#> [[6]]
#> <ResamplingSubsampling> with 30 iterations
#> * Instantiated: TRUE
#> * Parameters: repeats=30, ratio=0.6667
#> 
#> [[7]]
#> <ResamplingCV> with 10 iterations
#> * Instantiated: TRUE
#> * Parameters: folds=10
#> 
#> [[8]]
#> <ResamplingSubsampling> with 30 iterations
#> * Instantiated: TRUE
#> * Parameters: repeats=30, ratio=0.6667
#> 
if (FALSE) {
benchmark(grid)
}