mlr3: Machine Learning in R - Next Generation

Efficient, object-oriented programming on the building blocks of machine learning. Provides 'R6' objects for tasks, learners, resamplings, and measures. The package is geared towards scalability and larger datasets by supporting parallelization and out-of-memory data-backends like databases. While 'mlr3' focuses on the core computational operations, add-on packages provide additional functionality.

Learn mlr3

Book on mlr3: https://mlr3book.mlr-org.com

Use cases and examples gallery: https://mlr3gallery.mlr-org.com

Cheat Sheets: https://github.com/mlr-org/mlr3cheatsheets

mlr3 extensions

Preprocessing and machine learning pipelines: mlr3pipelines

Analysis of benchmark experiments: mlr3benchmark

More classification and regression tasks: mlr3data

Connector to OpenML: mlr3oml

Solid selection of good classification and regression learners: mlr3learners

Even more learners: https://github.com/mlr-org/mlr3extralearners

Tuning of hyperparameters: mlr3tuning

Hyperband tuner: mlr3hyperband

Visualizations for many mlr3 objects: mlr3viz

Survival analysis and probabilistic regression: mlr3proba

Cluster analysis: mlr3cluster

Feature selection filters: mlr3filters

Feature selection wrappers: mlr3fselect

Interface to real (out-of-memory) data bases: mlr3db

Performance measures as plain functions: mlr3measures

Resampling methods for spatiotemporal data: mlr3spatiotempcv

Data storage and prediction support for spatial objects: mlr3spatial

Suggested packages

Parallelization framework: future

Progress bars: progressr

Encapsulated evaluation: evaluate, callr (external process)

Package Options

"mlr3.exec_random": Randomize the order of execution in resample() and benchmark() during parallelization with future. Defaults to TRUE. Note that this does not affect the order of results.

"mlr3.exec_chunk_size": Number of iterations to perform in a single future::future() during parallelization with future. Defaults to 1.

"mlr3.exec_chunk_bins": Number of bins to split the iterations into. If set, "mlr3.exec_chunk_size" is ignored.

"mlr3.debug": If set to TRUE, parallelization via future is disabled to simplify debugging and provide more concise tracebacks. Note that results computed in debug mode use a different seeding mechanism and are not reproducible.

"mlr3.allow_utf8_names": If set to TRUE, checks on the feature names are relaxed, allowing non-ascii characters in column names. This is an experimental and temporal option to pave the way for text analysis, and will likely be removed in a future version of the package. analysis.

"mlr3.warn_version_mismatch": Set to FALSE to silence warnings raised during predict if a learner has been trained with a different version version of mlr3.

References

Lang M, Binder M, Richter J, Schratz P, Pfisterer F, Coors S, Au Q, Casalicchio G, Kotthoff L, Bischl B (2019). “mlr3: A modern object-oriented machine learning framework in R.” Journal of Open Source Software. doi:10.21105/joss.01903 , https://joss.theoj.org/papers/10.21105/joss.01903.

See also

Author

Maintainer: Marc Becker marcbecker@posteo.de (ORCID)

Authors:

Michel Lang michellang@gmail.com (ORCID)

Bernd Bischl bernd_bischl@gmx.net (ORCID)

Jakob Richter jakob1richter@gmail.com (ORCID)

Patrick Schratz patrick.schratz@gmail.com (ORCID)

Martin Binder mlr.developer@mb706.com

Florian Pfisterer pfistererf@googlemail.com (ORCID)

Raphael Sonabend raphaelsonabend@gmail.com (ORCID)

Sebastian Fischer sebf.fischer@gmail.com (ORCID)

Other contributors:

Giuseppe Casalicchio giuseppe.casalicchio@stat.uni-muenchen.de (ORCID) [contributor]

Stefan Coors mail@stefancoors.de (ORCID) [contributor]

Quay Au quayau@gmail.com (ORCID) [contributor]

Lennart Schneider lennart.sch@web.de (ORCID) [contributor]

Lona Koers lona.koers@gmail.com [contributor]