mlr3: Machine Learning in R - Next Generation

Efficient, object-oriented programming on the building blocks of machine learning. Provides 'R6' objects for tasks, learners, resamplings, and measures. The package is geared towards scalability and larger datasets by supporting parallelization and out-of-memory data-backends like databases. While 'mlr3' focuses on the core computational operations, add-on packages provide additional functionality.

Additional resources

Book on mlr3: https://mlr3book.mlr-org.com

Use cases and examples: https://mlr3gallery.mlr-org.com

More classification and regression learners: mlr3learners

Preprocessing and machine learning pipelines: mlr3pipelines

Tuning of hyperparameters: mlr3tuning

Visualizations for many mlr3 objects: mlr3viz

Survival analysis and probabilistic regression: mlr3proba

Feature selection filters: mlr3filters

Interface to real (out-of-memory) data bases: mlr3db

Performance measures as plain functions: mlr3measures

Parallelization framework: future

Progress bars: progressr

References

Lang M, Binder M, Richter J, Schratz P, Pfisterer F, Coors S, Au Q, Casalicchio G, Kotthoff L, Bischl B (2019). “mlr3: A modern object-oriented machine learning framework in R.” Journal of Open Source Software. doi: 10.21105/joss.01903 , https://joss.theoj.org/papers/10.21105/joss.01903.

See also