Assertion for mlr3 Objects

Functions intended to be used in packages extending mlr3. Most assertion functions ensure the right class attrbiture, and optionally additional properties. Additionally, the following compound assertions are implemented:

assert_learnable(task, learner)
(Task, Learner) -> NULL
Checks if the learner is applicable to the task. This includes type checks on the type, the feature types, and properties.

If an assertion fails, an exception is raised. Otherwise, the input object is returned invisibly.

assert_backend(b, .var.name = vname(b))

assert_task(
  task,
  task_type = NULL,
  feature_types = NULL,
  task_properties = NULL,
  .var.name = vname(task)
)

assert_tasks(
  tasks,
  task_type = NULL,
  feature_types = NULL,
  task_properties = NULL,
  .var.name = vname(tasks)
)

assert_learner(
  learner,
  task = NULL,
  properties = character(),
  .var.name = vname(learner)
)

assert_learners(
  learners,
  task = NULL,
  properties = character(),
  .var.name = vname(learners)
)

assert_learnable(task, learner)

assert_measure(
  measure,
  task = NULL,
  learner = NULL,
  .var.name = vname(measure)
)

assert_measures(
  measures,
  task = NULL,
  learner = NULL,
  .var.name = vname(measures)
)

assert_resampling(
  resampling,
  instantiated = NULL,
  .var.name = vname(resampling)
)

assert_resamplings(
  resamplings,
  instantiated = NULL,
  .var.name = vname(resamplings)
)

assert_prediction(prediction, .var.name = vname(prediction))

assert_resample_result(rr, .var.name = vname(rr))

assert_benchmark_result(bmr, .var.name = vname(bmr))

assert_row_ids(row_ids, null.ok = FALSE, .var.name = vname(row_ids))

Arguments

b
(DataBackend).

task
(Task).

feature_types
(character())
Feature types the learner operates on. Must be a subset of mlr_reflections$task_feature_types.

task_properties
(character())
Set of required task properties.

tasks
(list of Task).

learner
(Learner).

learners
(list of Learner).

measure
(Measure).

measures
(list of Measure).

resampling
(Resampling).

resamplings
(list of Resampling).

prediction
(Prediction).

bmr
(BenchmarkResult).

row_ids
(numeric()).

resample_result
(ResampleResult).