Dictionary of Learners

A simple mlr3misc::Dictionary storing objects of class Learner. Each learner has an associated help page, see mlr_learners_[id].

This dictionary can get populated with additional learners by add-on packages. For an opinionated set of solid classification and regression learners, install and load the mlr3learners package. More learners are connected via https://github.com/mlr-org/mlr3extralearners.

For a more convenient way to retrieve and construct learners, see lrn()/lrns().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "task_type", "feature_types", "packages", "properties", and "predict_types" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

See also

Examples

as.data.table(mlr_learners)
#>                    key                              label task_type
#> 1:       classif.debug   Debug Learner for Classification   classif
#> 2: classif.featureless Featureless Classification Learner   classif
#> 3:       classif.rpart                Classification Tree   classif
#> 4:          regr.debug       Debug Learner for Regression      regr
#> 5:    regr.featureless     Featureless Regression Learner      regr
#> 6:          regr.rpart                    Regression Tree      regr
#>                                           feature_types   packages
#> 1:     logical,integer,numeric,character,factor,ordered       mlr3
#> 2: logical,integer,numeric,character,factor,ordered,...       mlr3
#> 3:               logical,integer,numeric,factor,ordered mlr3,rpart
#> 4:     logical,integer,numeric,character,factor,ordered       mlr3
#> 5: logical,integer,numeric,character,factor,ordered,... mlr3,stats
#> 6:               logical,integer,numeric,factor,ordered mlr3,rpart
#>                                                               properties
#> 1:                         hotstart_forward,missings,multiclass,twoclass
#> 2: featureless,importance,missings,multiclass,selected_features,twoclass
#> 3:     importance,missings,multiclass,selected_features,twoclass,weights
#> 4:                                                              missings
#> 5:                     featureless,importance,missings,selected_features
#> 6:                         importance,missings,selected_features,weights
#>    predict_types
#> 1: response,prob
#> 2: response,prob
#> 3: response,prob
#> 4:   response,se
#> 5:   response,se
#> 6:      response
mlr_learners$get("classif.featureless")
#> <LearnerClassifFeatureless:classif.featureless>: Featureless Classification Learner
#> * Model: -
#> * Parameters: method=mode
#> * Packages: mlr3
#> * Predict Type: response
#> * Feature Types: logical, integer, numeric, character, factor, ordered,
#>   POSIXct
#> * Properties: featureless, importance, missings, multiclass,
#>   selected_features, twoclass
lrn("classif.rpart")
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: -
#> * Parameters: xval=0
#> * Packages: mlr3, rpart
#> * Predict Type: response
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights