Featureless Classification Learner

A simple LearnerClassif which only analyzes the labels during train, ignoring all features. Hyperparameter method determines the mode of operation during prediction:

mode:

Predicts the most frequent label. If there are two or more labels tied, randomly selects one per prediction.

sample:

Randomly predict a label uniformly.

weighted.sample:

Randomly predict a label, with probability estimated from the training distribution.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.featureless")
lrn("classif.featureless")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”, “POSIXct”

Required Packages: mlr3

Parameters

Id	Type	Default	Levels

method	character	mode	mode, sample, weighted.sample

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifFeatureless

Methods

Public methods

LearnerClassifFeatureless$new()

LearnerClassifFeatureless$importance()

LearnerClassifFeatureless$selected_features()

LearnerClassifFeatureless$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifFeatureless$new()

Method importance()

All features have a score of 0 for this learner.

Usage

LearnerClassifFeatureless$importance()

Returns

Named numeric().

Method selected_features()

Selected features are always the empty set for this learner.

Usage

LearnerClassifFeatureless$selected_features()

Returns

character(0).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifFeatureless$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.