Regression Learner for Debugging

A simple LearnerRegr used primarily in the unit tests and for debugging purposes. If no hyperparameter is set, it simply constantly predicts the mean value of the training data. The following hyperparameters trigger the following actions:

predict_missing:

Ratio of predictions which will be NA.

predict_missing_type:

To to encode missingness. “na” will insert NA values, “omit” will just return fewer predictions than requested.

save_tasks:

Saves input task in model slot during training and prediction.

threads:

Number of threads to use. Has no effect.

x:

Numeric tuning parameter. Has no effect.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.debug")
lrn("regr.debug")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”

Required Packages: -

Parameters

Id	Type	Default	Levels	Range

predict_missing	numeric	0		\([0, 1]\)

predict_missing_type	character	na	na, omit	\((-\infty, \infty)\)

save_tasks	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

threads	integer	-		\([1, \infty)\)

x	numeric	-		\([0, 1]\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrDebug

Methods

Public methods

LearnerRegrDebug$new()

LearnerRegrDebug$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrDebug$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrDebug$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = tsk("mtcars")
learner = lrn("regr.debug", save_tasks = TRUE)
learner$train(task, row_ids = 1:20)
prediction = learner$predict(task, row_ids = 21:32)

learner$model$task_train

#> <TaskRegr:mtcars> (20 x 11)
#> * Target: mpg
#> * Properties: -
#> * Features (10):
#>   - dbl (10): am, carb, cyl, disp, drat, gear, hp, qsec, vs, wt

learner$model$task_predict

#> <TaskRegr:mtcars> (12 x 11)
#> * Target: mpg
#> * Properties: -
#> * Features (10):
#>   - dbl (10): am, carb, cyl, disp, drat, gear, hp, qsec, vs, wt