Featureless Regression Learner

A simple LearnerRegr which only analyses the response during train, ignoring all features. If hyperparameter robust is FALSE (default), constantly predicts mean(y) as response and sd(y) as standard error. If robust is TRUE, median() and mad() are used instead of mean() and sd(), respectively.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.featureless")
lrn("regr.featureless")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”, “POSIXct”

Required Packages: 'stats'

Parameters

Id	Type	Default	Range	Levels

robust	logical	TRUE	\((-\infty, \infty)\)	TRUE, FALSE

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrFeatureless

Methods

Public methods

LearnerRegrFeatureless$new()

LearnerRegrFeatureless$importance()

LearnerRegrFeatureless$selected_features()

LearnerRegrFeatureless$clone()

Inherited methods

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrFeatureless$new()

Method importance()

All features have a score of 0 for this learner.

Usage

LearnerRegrFeatureless$importance()

Returns

Named numeric().

Method selected_features()

Selected features are always the empty set for this learner.

Usage

LearnerRegrFeatureless$selected_features()

Returns

character(0).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrFeatureless$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.