Dictionary of Performance Measures
Source: R/mlr_measures.R

A simple Dictionary storing objects of class Measure. Each measure has an associated help page, see mlr_measures_[id].

Format

R6::R6Class object.

Methods

get(key, ...)
(character(1), ...) -> any
Retrieves object with key key from the dictionary.

mget(keys, ...)
(character(), ...) -> named list()
Retrieves objects with keys keys from the dictionary, returns them in a list named with keys.

has(keys)
character() -> logical()
Returns a logical vector with TRUE at its i-th position, if the i-th key exists.

keys(pattern)
character(1) -> character()
Returns all keys which comply to the regular expression pattern.

add(key, value)
(character(1), any) -> self
Adds object value to the dictionary with key key, potentially overwriting a previously stored value.

remove(key)
character() -> self
Removes object with key key from the dictionary.

S3 methods

as.data.table(dict)
Dictionary -> data.table::data.table()
Returns a data.table() with fields "key", "task_type", "predict_type", and "packages" as columns.

See also

Examples

as.data.table(mlr_measures)

#>             key task_type predict_type packages
#> 1:  classif.acc   classif     response  Metrics
#> 2:  classif.auc   classif         prob  Metrics
#> 3: classif.mmce   classif     response  Metrics
#> 4:     regr.mse      regr     response  Metrics
#> 5:    time_both      <NA>     response         
#> 6: time_predict      <NA>     response         
#> 7:   time_train      <NA>     response         

mlr_measures$get("classif.mmce")

#> <MeasureClassifMMCE:classif.mmce>
#> Packages: Metrics
#> Range: [0, 1]
#> Minimize: TRUE
#> Predict type: response