Dictionary of Performance Measures

A simple mlr3misc::Dictionary storing objects of class Measure. Each measure has an associated help page, see mlr_measures_[id].

This dictionary can get populated with additional measures by add-on packages. E.g., mlr3proba adds survival measures and mlr3cluster adds cluster analysis measures.

For a more convenient way to retrieve and construct measures, see msr()/msrs().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "task_type", "predict_type", and "packages" as columns.

See also

Examples

as.data.table(mlr_measures)
#>                     key task_type          packages predict_type
#>  1:                 aic      <NA>              mlr3     response
#>  2:                 bic      <NA>              mlr3     response
#>  3:         classif.acc   classif mlr3,mlr3measures     response
#>  4:         classif.auc   classif mlr3,mlr3measures         prob
#>  5:        classif.bacc   classif mlr3,mlr3measures     response
#>  6:      classif.bbrier   classif mlr3,mlr3measures         prob
#>  7:          classif.ce   classif mlr3,mlr3measures     response
#>  8:       classif.costs   classif              mlr3     response
#>  9:         classif.dor   classif mlr3,mlr3measures     response
#> 10:       classif.fbeta   classif mlr3,mlr3measures     response
#> 11:         classif.fdr   classif mlr3,mlr3measures     response
#> 12:          classif.fn   classif mlr3,mlr3measures     response
#> 13:         classif.fnr   classif mlr3,mlr3measures     response
#> 14:        classif.fomr   classif mlr3,mlr3measures     response
#> 15:          classif.fp   classif mlr3,mlr3measures     response
#> 16:         classif.fpr   classif mlr3,mlr3measures     response
#> 17:     classif.logloss   classif mlr3,mlr3measures         prob
#> 18:      classif.mbrier   classif mlr3,mlr3measures         prob
#> 19:         classif.mcc   classif mlr3,mlr3measures     response
#> 20:         classif.npv   classif mlr3,mlr3measures     response
#> 21:         classif.ppv   classif mlr3,mlr3measures     response
#> 22:       classif.prauc   classif mlr3,mlr3measures         prob
#> 23:   classif.precision   classif mlr3,mlr3measures     response
#> 24:      classif.recall   classif mlr3,mlr3measures     response
#> 25: classif.sensitivity   classif mlr3,mlr3measures     response
#> 26: classif.specificity   classif mlr3,mlr3measures     response
#> 27:          classif.tn   classif mlr3,mlr3measures     response
#> 28:         classif.tnr   classif mlr3,mlr3measures     response
#> 29:          classif.tp   classif mlr3,mlr3measures     response
#> 30:         classif.tpr   classif mlr3,mlr3measures     response
#> 31:               debug      <NA>              mlr3     response
#> 32:           oob_error      <NA>              mlr3     response
#> 33:           regr.bias      regr mlr3,mlr3measures     response
#> 34:           regr.ktau      regr mlr3,mlr3measures     response
#> 35:            regr.mae      regr mlr3,mlr3measures     response
#> 36:           regr.mape      regr mlr3,mlr3measures     response
#> 37:          regr.maxae      regr mlr3,mlr3measures     response
#> 38:          regr.medae      regr mlr3,mlr3measures     response
#> 39:          regr.medse      regr mlr3,mlr3measures     response
#> 40:            regr.mse      regr mlr3,mlr3measures     response
#> 41:           regr.msle      regr mlr3,mlr3measures     response
#> 42:          regr.pbias      regr mlr3,mlr3measures     response
#> 43:            regr.rae      regr mlr3,mlr3measures     response
#> 44:           regr.rmse      regr mlr3,mlr3measures     response
#> 45:          regr.rmsle      regr mlr3,mlr3measures     response
#> 46:           regr.rrse      regr mlr3,mlr3measures     response
#> 47:            regr.rse      regr mlr3,mlr3measures     response
#> 48:            regr.rsq      regr mlr3,mlr3measures     response
#> 49:            regr.sae      regr mlr3,mlr3measures     response
#> 50:          regr.smape      regr mlr3,mlr3measures     response
#> 51:           regr.srho      regr mlr3,mlr3measures     response
#> 52:            regr.sse      regr mlr3,mlr3measures     response
#> 53:   selected_features      <NA>              mlr3     response
#> 54:         sim.jaccard      <NA> mlr3,mlr3measures     response
#> 55:             sim.phi      <NA> mlr3,mlr3measures     response
#> 56:           time_both      <NA>              mlr3     response
#> 57:        time_predict      <NA>              mlr3     response
#> 58:          time_train      <NA>              mlr3     response
#>                     key task_type          packages predict_type
#>     task_properties
#>  1:                
#>  2:                
#>  3:                
#>  4:        twoclass
#>  5:                
#>  6:        twoclass
#>  7:                
#>  8:                
#>  9:        twoclass
#> 10:        twoclass
#> 11:        twoclass
#> 12:        twoclass
#> 13:        twoclass
#> 14:        twoclass
#> 15:        twoclass
#> 16:        twoclass
#> 17:                
#> 18:                
#> 19:        twoclass
#> 20:        twoclass
#> 21:        twoclass
#> 22:        twoclass
#> 23:        twoclass
#> 24:        twoclass
#> 25:        twoclass
#> 26:        twoclass
#> 27:        twoclass
#> 28:        twoclass
#> 29:        twoclass
#> 30:        twoclass
#> 31:                
#> 32:                
#> 33:                
#> 34:                
#> 35:                
#> 36:                
#> 37:                
#> 38:                
#> 39:                
#> 40:                
#> 41:                
#> 42:                
#> 43:                
#> 44:                
#> 45:                
#> 46:                
#> 47:                
#> 48:                
#> 49:                
#> 50:                
#> 51:                
#> 52:                
#> 53:                
#> 54:                
#> 55:                
#> 56:                
#> 57:                
#> 58:                
#>     task_properties
mlr_measures$get("classif.ce")
#> <MeasureClassifSimple:classif.ce>
#> * Packages: mlr3, mlr3measures
#> * Range: [0, 1]
#> * Minimize: TRUE
#> * Average: macro
#> * Parameters: list()
#> * Properties: -
#> * Predict type: response
msr("regr.mse")
#> <MeasureRegrSimple:regr.mse>
#> * Packages: mlr3, mlr3measures
#> * Range: [0, Inf]
#> * Minimize: TRUE
#> * Average: macro
#> * Parameters: list()
#> * Properties: -
#> * Predict type: response