Dictionary of Performance Measures

A simple mlr3misc::Dictionary storing objects of class Measure. Each measure has an associated help page, see mlr_measures_[id].

This dictionary can get populated with additional measures by add-on packages.

For a more convenient way to retrieve and construct measures, see msr()/msrs().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "task_type", "predict_type", and "packages" as columns.

See also

Examples

as.data.table(mlr_measures)

#>                     key task_type     packages predict_type task_properties
#>  1:         classif.acc   classif mlr3measures     response                
#>  2:         classif.auc   classif mlr3measures         prob        twoclass
#>  3:        classif.bacc   classif mlr3measures     response                
#>  4:      classif.bbrier   classif mlr3measures         prob        twoclass
#>  5:          classif.ce   classif mlr3measures     response                
#>  6:       classif.costs   classif                  response                
#>  7:         classif.dor   classif mlr3measures     response        twoclass
#>  8:       classif.fbeta   classif mlr3measures     response        twoclass
#>  9:         classif.fdr   classif mlr3measures     response        twoclass
#> 10:          classif.fn   classif mlr3measures     response        twoclass
#> 11:         classif.fnr   classif mlr3measures     response        twoclass
#> 12:        classif.fomr   classif mlr3measures     response        twoclass
#> 13:          classif.fp   classif mlr3measures     response        twoclass
#> 14:         classif.fpr   classif mlr3measures     response        twoclass
#> 15:     classif.logloss   classif mlr3measures         prob                
#> 16:      classif.mbrier   classif mlr3measures         prob                
#> 17:         classif.mcc   classif mlr3measures     response        twoclass
#> 18:         classif.npv   classif mlr3measures     response        twoclass
#> 19:         classif.ppv   classif mlr3measures     response        twoclass
#> 20:   classif.precision   classif mlr3measures     response        twoclass
#> 21:      classif.recall   classif mlr3measures     response        twoclass
#> 22: classif.sensitivity   classif mlr3measures     response        twoclass
#> 23: classif.specificity   classif mlr3measures     response        twoclass
#> 24:          classif.tn   classif mlr3measures     response        twoclass
#> 25:         classif.tnr   classif mlr3measures     response        twoclass
#> 26:          classif.tp   classif mlr3measures     response        twoclass
#> 27:         classif.tpr   classif mlr3measures     response        twoclass
#> 28:               debug      <NA>                  response                
#> 29:           oob_error      <NA>                  response                
#> 30:           regr.bias      regr mlr3measures     response                
#> 31:           regr.ktau      regr mlr3measures     response                
#> 32:            regr.mae      regr mlr3measures     response                
#> 33:           regr.mape      regr mlr3measures     response                
#> 34:          regr.maxae      regr mlr3measures     response                
#> 35:          regr.medae      regr mlr3measures     response                
#> 36:          regr.medse      regr mlr3measures     response                
#> 37:            regr.mse      regr mlr3measures     response                
#> 38:           regr.msle      regr mlr3measures     response                
#> 39:          regr.pbias      regr mlr3measures     response                
#> 40:            regr.rae      regr mlr3measures     response                
#> 41:           regr.rmse      regr mlr3measures     response                
#> 42:          regr.rmsle      regr mlr3measures     response                
#> 43:           regr.rrse      regr mlr3measures     response                
#> 44:            regr.rse      regr mlr3measures     response                
#> 45:            regr.rsq      regr mlr3measures     response                
#> 46:            regr.sae      regr mlr3measures     response                
#> 47:          regr.smape      regr mlr3measures     response                
#> 48:           regr.srho      regr mlr3measures     response                
#> 49:            regr.sse      regr mlr3measures     response                
#> 50:   selected_features      <NA>                  response                
#> 51:           time_both      <NA>                  response                
#> 52:        time_predict      <NA>                  response                
#> 53:          time_train      <NA>                  response                
#>                     key task_type     packages predict_type task_properties

mlr_measures$get("classif.ce")

#> <MeasureClassifSimple:classif.ce>
#> * Packages: mlr3measures
#> * Range: [0, 1]
#> * Minimize: TRUE
#> * Properties: -
#> * Predict type: response

msr("regr.mse")

#> <MeasureRegrSimple:regr.mse>
#> * Packages: mlr3measures
#> * Range: [0, Inf]
#> * Minimize: TRUE
#> * Properties: -
#> * Predict type: response