Dictionary of Performance Measures

A simple mlr3misc::Dictionary storing objects of class Measure. Each measure has an associated help page, see mlr_measures_[id].

This dictionary can get populated with additional measures by add-on packages. E.g., mlr3proba adds survival measures and mlr3cluster adds cluster analysis measures.

For a more convenient way to retrieve and construct measures, see msr()/msrs().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "task_type", "packages", "predict_type", and "task_properties" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

See also

Examples

as.data.table(mlr_measures)
#>                     key                                               label
#>  1:                 aic                        Akaika Information Criterion
#>  2:                 bic                      Bayesian Information Criterion
#>  3:         classif.acc                             Classification Accuracy
#>  4:         classif.auc                            Area Under the ROC Curve
#>  5:        classif.bacc                                   Balanced Accuracy
#>  6:      classif.bbrier                                  Binary Brier Score
#>  7:          classif.ce                                Classification Error
#>  8:       classif.costs                       Cost-sensitive Classification
#>  9:         classif.dor                               Diagnostic Odds Ratio
#> 10:       classif.fbeta                                        F-beta score
#> 11:         classif.fdr                                False Discovery Rate
#> 12:          classif.fn                                     False Negatives
#> 13:         classif.fnr                                 False Negative Rate
#> 14:        classif.fomr                                 False Omission Rate
#> 15:          classif.fp                                     False Positives
#> 16:         classif.fpr                                 False Positive Rate
#> 17:     classif.logloss                                            Log Loss
#> 18:   classif.mauc_au1p             Weighted average 1 vs. 1 multiclass AUC
#> 19:   classif.mauc_au1u                      Average 1 vs. 1 multiclass AUC
#> 20:   classif.mauc_aunp          Weighted average 1 vs. rest multiclass AUC
#> 21:   classif.mauc_aunu                   Average 1 vs. rest multiclass AUC
#> 22:      classif.mbrier                              Multiclass Brier Score
#> 23:         classif.mcc                    Matthews Correlation Coefficient
#> 24:         classif.npv                           Negative Predictive Value
#> 25:         classif.ppv                           Positive Predictive Value
#> 26:       classif.prauc                              Precision-Recall Curve
#> 27:   classif.precision                                           Precision
#> 28:      classif.recall                                              Recall
#> 29: classif.sensitivity                                         Sensitivity
#> 30: classif.specificity                                         Specificity
#> 31:          classif.tn                                      True Negatives
#> 32:         classif.tnr                                  True Negative Rate
#> 33:          classif.tp                                      True Positives
#> 34:         classif.tpr                                  True Positive Rate
#> 35:               debug                        Debug Classification Measure
#> 36:           oob_error                                    Out-of-bag Error
#> 37:           regr.bias                                                Bias
#> 38:           regr.ktau                                       Kendall's tau
#> 39:            regr.mae                                 Mean Absolute Error
#> 40:           regr.mape                         Mean Absolute Percent Error
#> 41:          regr.maxae                                  Max Absolute Error
#> 42:          regr.medae                               Median Absolute Error
#> 43:          regr.medse                                Median Squared Error
#> 44:            regr.mse                                  Mean Squared Error
#> 45:           regr.msle                              Mean Squared Log Error
#> 46:          regr.pbias                                        Percent Bias
#> 47:            regr.rae                             Relative Absolute Error
#> 48:           regr.rmse                             Root Mean Squared Error
#> 49:          regr.rmsle                         Root Mean Squared Log Error
#> 50:           regr.rrse                         Root Relative Squared Error
#> 51:            regr.rse                              Relative Squared Error
#> 52:            regr.rsq                                           R Squared
#> 53:            regr.sae                              Sum of Absolute Errors
#> 54:          regr.smape               Symmetric Mean Absolute Percent Error
#> 55:           regr.srho                                      Spearman's rho
#> 56:            regr.sse                               Sum of Squared Errors
#> 57:   selected_features Absolute or Relative Frequency of Selected Features
#> 58:         sim.jaccard                            Jaccard Similarity Index
#> 59:             sim.phi                          Phi Coefficient Similarity
#> 60:           time_both                                        Elapsed Time
#> 61:        time_predict                                        Elapsed Time
#> 62:          time_train                                        Elapsed Time
#>                     key                                               label
#>     task_type          packages predict_type task_properties
#>  1:      <NA>              mlr3     response                
#>  2:      <NA>              mlr3     response                
#>  3:   classif mlr3,mlr3measures     response                
#>  4:   classif mlr3,mlr3measures         prob        twoclass
#>  5:   classif mlr3,mlr3measures     response                
#>  6:   classif mlr3,mlr3measures         prob        twoclass
#>  7:   classif mlr3,mlr3measures     response                
#>  8:   classif              mlr3     response                
#>  9:   classif mlr3,mlr3measures     response        twoclass
#> 10:   classif mlr3,mlr3measures     response        twoclass
#> 11:   classif mlr3,mlr3measures     response        twoclass
#> 12:   classif mlr3,mlr3measures     response        twoclass
#> 13:   classif mlr3,mlr3measures     response        twoclass
#> 14:   classif mlr3,mlr3measures     response        twoclass
#> 15:   classif mlr3,mlr3measures     response        twoclass
#> 16:   classif mlr3,mlr3measures     response        twoclass
#> 17:   classif mlr3,mlr3measures         prob                
#> 18:   classif mlr3,mlr3measures         prob                
#> 19:   classif mlr3,mlr3measures         prob                
#> 20:   classif mlr3,mlr3measures         prob                
#> 21:   classif mlr3,mlr3measures         prob                
#> 22:   classif mlr3,mlr3measures         prob                
#> 23:   classif mlr3,mlr3measures     response        twoclass
#> 24:   classif mlr3,mlr3measures     response        twoclass
#> 25:   classif mlr3,mlr3measures     response        twoclass
#> 26:   classif mlr3,mlr3measures         prob        twoclass
#> 27:   classif mlr3,mlr3measures     response        twoclass
#> 28:   classif mlr3,mlr3measures     response        twoclass
#> 29:   classif mlr3,mlr3measures     response        twoclass
#> 30:   classif mlr3,mlr3measures     response        twoclass
#> 31:   classif mlr3,mlr3measures     response        twoclass
#> 32:   classif mlr3,mlr3measures     response        twoclass
#> 33:   classif mlr3,mlr3measures     response        twoclass
#> 34:   classif mlr3,mlr3measures     response        twoclass
#> 35:      <NA>              mlr3     response                
#> 36:      <NA>              mlr3     response                
#> 37:      regr mlr3,mlr3measures     response                
#> 38:      regr mlr3,mlr3measures     response                
#> 39:      regr mlr3,mlr3measures     response                
#> 40:      regr mlr3,mlr3measures     response                
#> 41:      regr mlr3,mlr3measures     response                
#> 42:      regr mlr3,mlr3measures     response                
#> 43:      regr mlr3,mlr3measures     response                
#> 44:      regr mlr3,mlr3measures     response                
#> 45:      regr mlr3,mlr3measures     response                
#> 46:      regr mlr3,mlr3measures     response                
#> 47:      regr mlr3,mlr3measures     response                
#> 48:      regr mlr3,mlr3measures     response                
#> 49:      regr mlr3,mlr3measures     response                
#> 50:      regr mlr3,mlr3measures     response                
#> 51:      regr mlr3,mlr3measures     response                
#> 52:      regr mlr3,mlr3measures     response                
#> 53:      regr mlr3,mlr3measures     response                
#> 54:      regr mlr3,mlr3measures     response                
#> 55:      regr mlr3,mlr3measures     response                
#> 56:      regr mlr3,mlr3measures     response                
#> 57:      <NA>              mlr3     response                
#> 58:      <NA> mlr3,mlr3measures     response                
#> 59:      <NA> mlr3,mlr3measures     response                
#> 60:      <NA>              mlr3         <NA>                
#> 61:      <NA>              mlr3         <NA>                
#> 62:      <NA>              mlr3         <NA>                
#>     task_type          packages predict_type task_properties
mlr_measures$get("classif.ce")
#> <MeasureClassifSimple:classif.ce>: Classification Error
#> * Packages: mlr3, mlr3measures
#> * Range: [0, 1]
#> * Minimize: TRUE
#> * Average: macro
#> * Parameters: list()
#> * Properties: -
#> * Predict type: response
msr("regr.mse")
#> <MeasureRegrSimple:regr.mse>: Mean Squared Error
#> * Packages: mlr3, mlr3measures
#> * Range: [0, Inf]
#> * Minimize: TRUE
#> * Average: macro
#> * Parameters: list()
#> * Properties: -
#> * Predict type: response