Cost-sensitive Classification Measure

Uses a cost matrix to create a classification measure. True labels must be arranged in columns, predicted labels must be arranged in rows. The cost matrix is stored as slot $costs.

For calculation of the score, the confusion matrix is multiplied element-wise with the cost matrix. The costs are then summed up (and potentially divided by the number of observations if normalize is set to TRUE).

This measure requires the Task during scoring to ensure that the rows and columns of the cost matrix are in the same order as in the confusion matrix.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("classif.costs")
msr("classif.costs")

Meta Information

Type: "classif"

Range: \([0, \infty)\)

Minimize: TRUE

Required prediction: 'response'

See also

Super classes

mlr3::Measure -> mlr3::MeasureClassif -> MeasureClassifCosts

Public fields

normalize

(logical(1))
Normalize the costs?

Active bindings

costs

(numeric matrix())
Matrix of costs (truth in columns, predicted response in rows).

Methods

Public methods

MeasureClassifCosts$new()

MeasureClassifCosts$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureClassifCosts$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureClassifCosts$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# get a cost sensitive task
task = tsk("german_credit")

# cost matrix as given on the UCI page of the german credit data set
# https://archive.ics.uci.edu/ml/datasets/statlog+(german+credit+data)
costs = matrix(c(0, 5, 1, 0), nrow = 2)
dimnames(costs) = list(truth = task$class_names, predicted = task$class_names)
print(costs)

#>       predicted
#> truth  good bad
#>   good    0   1
#>   bad     5   0

# mlr3 needs truth in columns, predictions in rows
costs = t(costs)

# create measure which calculates the absolute costs
m = msr("classif.costs", id = "german_credit_costs", costs = costs, normalize = FALSE)

# fit models and calculate costs
learner = lrn("classif.rpart")
rr = resample(task, learner, rsmp("cv", folds = 3))

#> INFO  [10:08:56.459] Applying learner 'classif.rpart' on task 'german_credit' (iter 1/3) 
#> INFO  [10:08:56.488] Applying learner 'classif.rpart' on task 'german_credit' (iter 2/3) 
#> INFO  [10:08:56.517] Applying learner 'classif.rpart' on task 'german_credit' (iter 3/3) 

rr$aggregate(m)

#> german_credit_costs 
#>            296.6667