Debug Measure

This measure returns the number of observations in the Prediction object. Its main purpose is debugging. The parameter na_ratio (numeric(1)) controls the ratio of scores which randomly are set to NA, between 0 (default) and 1.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

mlr_measures$get("debug")
msr("debug")

Meta Information

Task type: “NA”

Range: \([0, \infty)\)

Minimize: NA

Average: macro

Required Prediction: “response”

Required Packages: -

Parameters

Id	Type	Default	Range

na_ratio	numeric	-	\([0, 1]\)

See also

Super class

mlr3::Measure -> MeasureDebug

Methods

Public methods

MeasureDebug$new()

MeasureDebug$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureDebug$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureDebug$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = tsk("wine")
learner = lrn("classif.featureless")
measure = msr("debug", na_ratio = 0.5)
rr = resample(task, learner, rsmp("cv", folds = 5))
rr$score(measure)

#>                 task task_id                         learner
#> 1: <TaskClassif[49]>    wine <LearnerClassifFeatureless[37]>
#> 2: <TaskClassif[49]>    wine <LearnerClassifFeatureless[37]>
#> 3: <TaskClassif[49]>    wine <LearnerClassifFeatureless[37]>
#> 4: <TaskClassif[49]>    wine <LearnerClassifFeatureless[37]>
#> 5: <TaskClassif[49]>    wine <LearnerClassifFeatureless[37]>
#>             learner_id         resampling resampling_id iteration
#> 1: classif.featureless <ResamplingCV[19]>            cv         1
#> 2: classif.featureless <ResamplingCV[19]>            cv         2
#> 3: classif.featureless <ResamplingCV[19]>            cv         3
#> 4: classif.featureless <ResamplingCV[19]>            cv         4
#> 5: classif.featureless <ResamplingCV[19]>            cv         5
#>                 prediction debug
#> 1: <PredictionClassif[20]>    NA
#> 2: <PredictionClassif[20]>    36
#> 3: <PredictionClassif[20]>    NA
#> 4: <PredictionClassif[20]>    35
#> 5: <PredictionClassif[20]>    NA