Out-of-bag Error Measure

Returns the out-of-bag error of the Learner for learners that support it (learners with property "oob_error"). Returns NA for unsupported learners.

Dictionary

This Measure can be instantiated via the dictionary mlr_measures or with the associated sugar function msr():

Meta Information

, * Task type: “NA”, * Range: \((-\infty, \infty)\), * Minimize: TRUE, * Average: macro, * Required Prediction: “response”, * Required Packages: mlr3

Parameters

Empty ParamSet

See also

Super class

mlr3::Measure -> MeasureOOBError

Methods

Public methods

MeasureOOBError$new()

MeasureOOBError$clone()

Inherited methods

mlr3::Measure$aggregate()

mlr3::Measure$format()

mlr3::Measure$help()

mlr3::Measure$print()

mlr3::Measure$score()

Method new()

Creates a new instance of this R6 class.

Usage

MeasureOOBError$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

MeasureOOBError$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.