Dictionary of Resampling Strategies

A simple mlr3misc::Dictionary storing objects of class Resampling. Each resampling has an associated help page, see mlr_resamplings_[id].

This dictionary can get populated with additional resampling strategies by add-on packages.

For a more convenient way to retrieve and construct resampling strategies, see rsmp()/rsmps().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with columns "key", "label", "params", and "iters". If objects is set to TRUE, the constructed objects are returned in the list column named object.

See also

Examples

as.data.table(mlr_resamplings)
#> Key: <key>
#>            key                         label        params iters
#>         <char>                        <char>        <list> <int>
#> 1:   bootstrap                     Bootstrap ratio,repeats    30
#> 2:      custom                 Custom Splits                  NA
#> 3:   custom_cv Custom Split Cross-Validation                  NA
#> 4:          cv              Cross-Validation         folds    10
#> 5:     holdout                       Holdout         ratio     1
#> 6:    insample           Insample Resampling                   1
#> 7:         loo                 Leave-One-Out                  NA
#> 8: repeated_cv     Repeated Cross-Validation folds,repeats   100
#> 9: subsampling                   Subsampling ratio,repeats    30
mlr_resamplings$get("cv")
#> 
#> ── <ResamplingCV> : Cross-Validation ───────────────────────────────────────────
#> • Iterations: 10
#> • Instantiated: FALSE
#> • Parameters: folds=10
rsmp("subsampling")
#> 
#> ── <ResamplingSubsampling> : Subsampling ───────────────────────────────────────
#> • Iterations: 30
#> • Instantiated: FALSE
#> • Parameters: ratio=0.6667, repeats=30