Cross-Validation Resampling

Splits data using a folds-folds (default: 10 folds) cross-validation.

Dictionary

This Resampling can be instantiated via the dictionary mlr_resamplings or with the associated sugar function rsmp():

mlr_resamplings$get("cv")
rsmp("cv")

Parameters

folds (integer(1))
Number of folds.

References

Bischl B, Mersmann O, Trautmann H, Weihs C (2012). “Resampling Methods for Meta-Model Validation with Recommendations for Evolutionary Computation.” Evolutionary Computation, 20(2), 249--275. doi: 10.1162/evco_a_00069 .

See also

Super class

mlr3::Resampling -> ResamplingCV

Active bindings

iters

(integer(1))
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingCV$new()

ResamplingCV$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$instantiate()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Creates a new instance of this R6 class.

Usage

ResamplingCV$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingCV$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a task with 10 observations
task = tsk("penguins")
task$filter(1:10)

# Instantiate Resampling
cv = rsmp("cv", folds = 3)
cv$instantiate(task)

# Individual sets:
cv$train_set(1)
#> [1]  7  8 10  2  3  4
cv$test_set(1)
#> [1] 1 5 6 9

# Disjunct sets:
intersect(cv$train_set(1), cv$test_set(1))
#> integer(0)

# Internal storage:
cv$instance # table
#>     row_id fold
#>  1:      1    1
#>  2:      5    1
#>  3:      6    1
#>  4:      9    1
#>  5:      7    2
#>  6:      8    2
#>  7:     10    2
#>  8:      2    3
#>  9:      3    3
#> 10:      4    3