Repeated Cross-Validation Resampling

Splits data repeats (default: 10) times using a folds-fold (default: 10) cross-validation.

The iteration counter translates to repeats blocks of folds cross-validations, i.e., the first folds iterations belong to a single cross-validation.

Iteration numbers can be translated into folds or repeats with provided methods.

Dictionary

This Resampling can be instantiated via the dictionary mlr_resamplings or with the associated sugar function rsmp():

mlr_resamplings$get("repeated_cv")
rsmp("repeated_cv")

Parameters

repeats (integer(1))
Number of repetitions.

folds (integer(1))
Number of folds.

References

Bischl B, Mersmann O, Trautmann H, Weihs C (2012). “Resampling Methods for Meta-Model Validation with Recommendations for Evolutionary Computation.” Evolutionary Computation, 20(2), 249--275. doi: 10.1162/evco_a_00069 .

See also

Super class

mlr3::Resampling -> ResamplingRepeatedCV

Active bindings

iters

(integer(1))
Returns the number of resampling iterations, depending on the values stored in the param_set.

Methods

Public methods

ResamplingRepeatedCV$new()

ResamplingRepeatedCV$folds()

ResamplingRepeatedCV$repeats()

ResamplingRepeatedCV$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$instantiate()

mlr3::Resampling$print()

mlr3::Resampling$test_set()

mlr3::Resampling$train_set()

Method new()

Creates a new instance of this R6 class.

Usage

ResamplingRepeatedCV$new()

Method folds()

Translates iteration numbers to fold numbers.

Usage

ResamplingRepeatedCV$folds(iters)

Arguments

iters

(integer())
Iteration number.

Returns

integer() of fold numbers.

Method repeats()

Translates iteration numbers to repetition numbers.

Usage

ResamplingRepeatedCV$repeats(iters)

Arguments

iters

(integer())
Iteration number.

Returns

integer() of repetition numbers.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRepeatedCV$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a task with 10 observations
task = tsk("iris")
task$filter(1:10)

# Instantiate Resampling
rrcv = rsmp("repeated_cv", repeats = 2, folds = 3)
rrcv$instantiate(task)
rrcv$iters

#> [1] 6

rrcv$folds(1:6)

#> [1] 1 2 1 2 1 2

rrcv$repeats(1:6)

#> [1] 1 1 1 2 2 2

# Individual sets:
rrcv$train_set(1)

#> [1] 2 3 8 4 5 7

rrcv$test_set(1)

#> [1]  1  6  9 10

intersect(rrcv$train_set(1), rrcv$test_set(1))

#> integer(0)

# Internal storage:
rrcv$instance # table

#>     row_id rep fold
#>  1:      1   1    1
#>  2:      2   1    2
#>  3:      3   1    2
#>  4:      4   1    3
#>  5:      5   1    3
#>  6:      6   1    1
#>  7:      7   1    3
#>  8:      8   1    2
#>  9:      9   1    1
#> 10:     10   1    1
#> 11:      1   2    1
#> 12:      2   2    3
#> 13:      3   2    1
#> 14:      4   2    3
#> 15:      5   2    1
#> 16:      6   2    2
#> 17:      7   2    3
#> 18:      8   2    1
#> 19:      9   2    2
#> 20:     10   2    2