Syntactic Sugar for Object Construction

Functions to retrieve objects, set hyperparameters and assign to fields in one go. Relies on mlr3misc::dictionary_sugar_get() to extract objects from the respective mlr3misc::Dictionary:

tsk() for a Task from mlr_tasks.

tsks() for a list of Tasks from mlr_tasks.

tgen() for a TaskGenerator from mlr_task_generators.

tgens() for a list of TaskGenerators from mlr_task_generators.

lrn() for a Learner from mlr_learners.

lrns() for a list of Learners from mlr_learners.

rsmp() for a Resampling from mlr_resamplings.

rsmps() for a list of Resamplings from mlr_resamplings.

msr() for a Measure from mlr_measures.

msrs() for a list of Measures from mlr_measures.

Helper function to configure the $validate field(s) of a Learner.

This is especially useful for learners such as AutoTuner of mlr3tuning or GraphLearner of mlr3pipelines which have multiple levels of $validate fields., where the $validate fields need to be configured on multiple levels.

Usage

tsk(.key, ...)

tsks(.keys, ...)

tgen(.key, ...)

tgens(.keys, ...)

lrn(.key, ...)

lrns(.keys, ...)

rsmp(.key, ...)

rsmps(.keys, ...)

msr(.key, ...)

msrs(.keys, ...)

set_validate(learner, validate, ...)

Arguments

.key

(character(1))
Key passed to the respective dictionary to retrieve the object.

...

(any)
Additional arguments.

.keys

(character())
Keys passed to the respective dictionary to retrieve multiple objects.

learner

(any)
The learner.

validate

(numeric(1), "predefined", "test", or NULL)
Which validation set to use.

Value

R6::R6Class object of the respective type, or a list of R6::R6Class objects for the plural versions.

Modified Learner

Examples

# penguins task with new id
tsk("penguins", id = "penguins2")
#> <TaskClassif:penguins2> (344 x 8): Palmer Penguins
#> * Target: species
#> * Properties: multiclass
#> * Features (7):
#>   - int (3): body_mass, flipper_length, year
#>   - dbl (2): bill_depth, bill_length
#>   - fct (2): island, sex

# classification tree with different hyperparameters
# and predict type set to predict probabilities
lrn("classif.rpart", cp = 0.1, predict_type = "prob")
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: -
#> * Parameters: cp=0.1, xval=0
#> * Packages: mlr3, rpart
#> * Predict Types:  response, [prob]
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

# multiple learners with predict type 'prob'
lrns(c("classif.featureless", "classif.rpart"), predict_type = "prob")
#> $classif.featureless
#> <LearnerClassifFeatureless:classif.featureless>: Featureless Classification Learner
#> * Model: -
#> * Parameters: method=mode
#> * Packages: mlr3
#> * Predict Types:  response, [prob]
#> * Feature Types: logical, integer, numeric, character, factor, ordered,
#>   POSIXct
#> * Properties: featureless, importance, missings, multiclass,
#>   selected_features, twoclass
#> 
#> $classif.rpart
#> <LearnerClassifRpart:classif.rpart>: Classification Tree
#> * Model: -
#> * Parameters: xval=0
#> * Packages: mlr3, rpart
#> * Predict Types:  response, [prob]
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#> 
learner = lrn("classif.debug")
set_validate(learner, 0.2)
learner$validate
#> [1] 0.2