Syntactic Sugar for Object Construction

Functions to retrieve objects, set hyperparameters and assign to fields in one go. Relies on mlr3misc::dictionary_sugar_get() to extract objects from the respective mlr3misc::Dictionary:

tsk() for a Task from mlr_tasks.

tsks() for a list of Tasks from mlr_tasks.

tgen() for a TaskGenerator from mlr_task_generators.

tgens() for a list of TaskGenerators from mlr_task_generators.

lrn() for a Learner from mlr_learners.

lrns() for a list of Learners from mlr_learners.

rsmp() for a Resampling from mlr_resamplings.

rsmps() for a list of Resamplings from mlr_resamplings.

msr() for a Measure from mlr_measures.

msrs() for a list of Measures from mlr_measures.

tsk(.key, ...)

tsks(.keys, ...)

tgen(.key, ...)

tgens(.keys, ...)

lrn(.key, ...)

lrns(.keys, ...)

rsmp(.key, ...)

rsmps(.keys, ...)

msr(.key, ...)

msrs(.keys, ...)

Arguments

.key
(character(1))
Key passed to the respective dictionary to retrieve the object.

...
(named list())
Named arguments passed to the constructor, to be set as parameters in the paradox::ParamSet, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.

.keys
(character())
Keys passed to the respective dictionary to retrieve multiple objects.

Value

R6::R6Class object of the respective type, or a list of R6::R6Class objects for the plural versions.

Examples

# iris task with new id
tsk("iris", id = "iris2")

#> <TaskClassif:iris2> (150 x 5)
#> * Target: Species
#> * Properties: multiclass
#> * Features (4):
#>   - dbl (4): Petal.Length, Petal.Width, Sepal.Length, Sepal.Width

# classification tree with different hyperparameters
# and predict type set to predict probabilities
lrn("classif.rpart", cp = 0.1, predict_type = "prob")

#> <LearnerClassifRpart:classif.rpart>
#> * Model: -
#> * Parameters: xval=0, cp=0.1
#> * Packages: rpart
#> * Predict Type: prob
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights

# multiple learners with predict type 'prob'
lrns(c("classif.featureless", "classif.rpart"), predict_type = "prob")

#> [[1]]
#> <LearnerClassifFeatureless:classif.featureless>
#> * Model: -
#> * Parameters: method=mode
#> * Packages: -
#> * Predict Type: prob
#> * Feature types: logical, integer, numeric, character, factor, ordered,
#>   POSIXct
#> * Properties: featureless, importance, missings, multiclass,
#>   selected_features, twoclass
#> 
#> [[2]]
#> <LearnerClassifRpart:classif.rpart>
#> * Model: -
#> * Parameters: xval=0
#> * Packages: rpart
#> * Predict Type: prob
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: importance, missings, multiclass, selected_features,
#>   twoclass, weights
#>