Dictionary of Task Generators

A simple mlr3misc::Dictionary storing objects of class TaskGenerator. Each task generator has an associated help page, see mlr_task_generators_[id].

This dictionary can get populated with additional task generators by add-on packages.

For a more convenient way to retrieve and construct task generators, see tgen().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key" and "packages" as columns.

See also

Examples

mlr_task_generators$get("smiley")

#> <TaskGeneratorSmiley>
#>   Inherits from: <TaskGenerator>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     generate: function (n) 
#>     id: smiley
#>     initialize: function () 
#>     man: mlr3::mlr_task_generators_smiley
#>     packages: mlbench
#>     param_set: ParamSet, R6
#>     task_type: classif
#>   Private:
#>     .generate: function (n) 

tgen("2dnormals")

#> <TaskGenerator2DNormals>
#>   Inherits from: <TaskGenerator>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     generate: function (n) 
#>     id: 2dnormals
#>     initialize: function () 
#>     man: mlr3::mlr_task_generators_2dnormals
#>     packages: mlbench
#>     param_set: ParamSet, R6
#>     task_type: classif
#>   Private:
#>     .generate: function (n)