Dictionary of Task Generators

A simple mlr3misc::Dictionary storing objects of class TaskGenerator. Each task generator has an associated help page, see mlr_task_generators_[id].

This dictionary can get populated with additional task generators by add-on packages.

For a more convenient way to retrieve and construct task generators, see tgen()/tgens().

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "task_type", "params", and "packages" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

See also

Examples

mlr_task_generators$get("smiley")
#> 
#> ── <TaskGeneratorSmiley> (smiley): Smiley Classification ───────────────────────
#> • Task type: classif
#> • Packages: mlr3 and mlbench
#> • Parameters: list()
#> • Manual: `?mlr3::mlr_task_generators_smiley()`
tgen("2dnormals")
#> 
#> ── <TaskGenerator2DNormals> (2dnormals): 2D Normals Classification ─────────────
#> • Task type: classif
#> • Packages: mlr3 and mlbench
#> • Parameters: list()
#> • Manual: `?mlr3::mlr_task_generators_2dnormals()`