Dictionary of Tasks
Source: R/mlr_tasks.R

A simple Dictionary storing objects of class Task.

Usage

See Dictionary.

See also

Examples

mlr_tasks$keys()

#> [1] "bh"      "iris"    "pima"    "sonar"   "spam"    "titanic" "zoo"    

as.data.table(mlr_tasks)
mlr_tasks$get("iris")

#> Task 'iris' of type classif (150 x 5)
#> Target: Species
#> Features: Sepal.Length, Sepal.Width, Petal.Length, Petal.Width
#> 
#> Public: all_classes, backend, cbind, class_n, class_names, clone,
#>   col_info, col_roles, data, feature_names, feature_types, filter,
#>   formula, hash, head, id, levels, measures, ncol, negative, nrow,
#>   overwrite, positive, properties, rbind, row_ids, row_roles, select,
#>   set_col_role, set_row_role, target_names, task_type, truth

head(mlr_tasks$get("iris")$data())

#>    Species Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1:  setosa          5.1         3.5          1.4         0.2
#> 2:  setosa          4.9         3.0          1.4         0.2
#> 3:  setosa          4.7         3.2          1.3         0.2
#> 4:  setosa          4.6         3.1          1.5         0.2
#> 5:  setosa          5.0         3.6          1.4         0.2
#> 6:  setosa          5.4         3.9          1.7         0.4

# Add a new task, based on a subset of iris:
data = iris
data$Species = factor(ifelse(data$Species == "setosa", "1", "0"))
b = as_data_backend(data)
task = TaskClassif$new("iris.binary", b, target = "Species", positive = "1")

# add to dictionary
mlr_tasks$add("iris.binary", task)

# list available tasks
mlr_tasks$keys()

#> [1] "bh"          "iris"        "iris.binary" "pima"        "sonar"      
#> [6] "spam"        "titanic"     "zoo"        

# retrieve from dictionary
mlr_tasks$get("iris.binary")

#> Task 'iris.binary' of type classif (150 x 5)
#> Target: Species
#> Features: Sepal.Length, Sepal.Width, Petal.Length, Petal.Width
#> 
#> Public: all_classes, backend, cbind, class_n, class_names, clone,
#>   col_info, col_roles, data, feature_names, feature_types, filter,
#>   formula, hash, head, id, levels, measures, ncol, negative, nrow,
#>   overwrite, positive, properties, rbind, row_ids, row_roles, select,
#>   set_col_role, set_row_role, target_names, task_type, truth

# remove task again
mlr_tasks$remove("iris.binary")