Resample a Learner on a Task
Source: R/resample.R

Runs a resampling (possibly in parallel).

resample(task, learner, resampling, measures = NULL, ctrl = list())

Arguments

task
:: (Task | character(1))
Object of type Task. Instead if a Task object, it is also possible to provide a key to retrieve a task from the mlr_tasks dictionary.

learner
:: (Learner | character(1))
Object of type Learner. Instead if a Learner object, it is also possible to provide a key to retrieve a learner from the mlr_learners dictionary.

resampling
:: (Resampling | character(1))
Object of type Resampling. Instead if a Resampling object, it is also possible to provide a key to retrieve a resampling from the mlr_resamplings dictionary.

measures
:: list of Measure
List of performance measures to calculate. Defaults to the measures specified in the Task task.

ctrl
:: named list()
Object to control experiment execution. See mlr_control() for details.

Value

ResampleResult.

Parallelization

This function can be parallelized with the future package. One jobs is one resampling iteration, and all jobs are forwarded to the future package together. To select a parallel backend, use future::plan().

Examples

task = mlr_tasks$get("iris")
learner = mlr_learners$get("classif.rpart")
resampling = mlr_resamplings$get("cv")

# explicitly instantiate the resampling for this task for reproduciblity
set.seed(123)
resampling$instantiate(task)

rr = resample(task, learner, resampling)
print(rr, digits = 2)

#> <ResampleResult> of learner 'iris' on task 'classif.rpart' with 10 iterations
#>     Measure Min. 1st Qu. Median Mean 3rd Qu. Max.   Sd
#>  classif.ce    0       0  0.033 0.06   0.067  0.2 0.08

# retrieve performance
rr$performance("classif.ce")

#>  [1] 0.00000000 0.20000000 0.20000000 0.06666667 0.06666667 0.00000000
#>  [7] 0.00000000 0.00000000 0.06666667 0.00000000

rr$aggregated

#> classif.ce 
#>       0.06 

# merged prediction object for all experiments
pred = rr$prediction
pred$confusion

#>             truth
#> response     setosa versicolor virginica
#>   setosa         50          0         0
#>   versicolor      0         46         5
#>   virginica       0          4        45

# Repeat resampling with featureless learner
rr.featureless = resample(task, "classif.featureless", resampling)

# Combine the ResampleResults into a BenchmarkResult
bmr = rr$combine(rr.featureless)
bmr$aggregated(objects = FALSE)

#>                hash task_id          learner_id resampling_id classif.ce
#> 1: b4875e7bc8ebb4cb    iris       classif.rpart            cv       0.06
#> 2: 19650135edd3897f    iris classif.featureless            cv       0.80